% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone data set}
\format{A data frame with 4177 rows and 9 variables.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Abalone}
}
\usage{
abalone
}
\description{
Abalone data set from the UCI Machine Learning Repository.
}
\keyword{datasets}
