% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{admit_columns.patch}
\alias{admit_columns.patch}
\title{Default implementation of \code{admit_columns} for column-wise patches.}
\usage{
\method{admit_columns}{patch}(patch, cols, cols_param_name = "cols")
}
\arguments{
\item{patch}{An elementary patch object.}

\item{cols}{A vector of integer column identifiers.}

\item{cols_param_name}{The name of the relevant column index parameter in the \code{patch}.}
}
\value{
This function is invoked for its side effect. It returns \code{NULL}
invisibly.
}
\description{
Default implementation of \code{admit_columns} for column-wise patches.
}
\keyword{internal}
