% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{admit_columns.patch_identity}
\alias{admit_columns.patch_identity}
\title{Identity patch implementation of \code{admit_columns}}
\usage{
\method{admit_columns}{patch_identity}(patch, cols, cols_param_name)
}
\arguments{
\item{patch}{An elementary patch object.}

\item{cols}{A vector of integer column identifiers.}

\item{cols_param_name}{The name of the relevant column index parameter in the \code{patch}.}
}
\value{
This function is invoked for its side effect, which is no effect
whatsoever in this case.
}
\description{
Identity patch implementation of \code{admit_columns}
}
