% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{admit_columns.patch_insert}
\alias{admit_columns.patch_insert}
\title{Insert patch implementation of \code{admit_columns}}
\usage{
\method{admit_columns}{patch_insert}(patch, cols,
  cols_param_name = "insertion_point")
}
\arguments{
\item{patch}{An elementary patch object.}

\item{cols}{A vector of integer column identifiers which are compatible with the
permutation in the given \code{patch}.}

\item{cols_param_name}{The name of the relevant column index parameter in the \code{patch}.}
}
\value{
This function is invoked for its side effect. The return value is
that of a tail call to the \code{\link{assign}} function, which is the
value assigned to the parameter named in the \code{cols_param_name} argument.
}
\description{
Insert patch implementation of \code{admit_columns}
}
\keyword{internal}
