% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{admit_columns}
\alias{admit_columns}
\title{Admit new column indices}
\usage{
admit_columns(patch, cols, cols_param_name)
}
\arguments{
\item{patch}{An elementary patch object.}

\item{cols}{A vector of integer column identifiers.}

\item{cols_param_name}{The name of the relevant column index parameter in the \code{patch}.}
}
\value{
This function is invoked for its side effect.
}
\description{
Modifies the index parameter in the given \code{patch} to admit the existence
of new columns at the positions specified by the \code{cols} argument.
}
\examples{
patch <- patch_permute(6:1)
patch
admit_columns(patch, cols = 3:4)
patch

}
