% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_bounds}
\alias{assert_bounds}
\title{Assert that a numeric vector is bounded by a min and/or max value}
\usage{
assert_bounds(x, exact = NULL, min = NULL, max = NULL, fn = NULL, label = NULL)
}
\arguments{
\item{x}{The numeric vector to check, or an object to pass to \code{fn}}

\item{exact}{Exact value \code{x} should have (optional)}

\item{min}{Minimum value \code{x} should have (optional)}

\item{max}{Maximum value \code{x} should have (optional)}

\item{fn}{Function to apply to x that returns the numeric vector to check (optional)}

\item{arg_name}{Optional name of the argument for more informative error messages,
defaults to the deparsed name of the object}
}
\value{
Invisibly returns the object if it passes the check, otherwise aborts
}
\description{
Assert that a numeric vector is bounded by a min and/or max value
}
