% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_class}
\alias{assert_class}
\title{Assert that an object belongs to a specified class}
\usage{
assert_class(x, class, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{The object to check}

\item{class}{A character vector of class names}

\item{arg_name}{Optional name of the argument for more informative error messages,
defaults to the deparsed name of the object}
}
\value{
Invisibly returns the object if it passes the check, otherwise aborts
}
\description{
Assert that an object belongs to a specified class
}
\examples{
x <- data.frame(a = 1)
assert_class(x, "data.frame")  # Passes
\dontrun{
assert_class(x, "matrix")      # Errors
assert_class(x, c("matrix", "list"))  # Passes (data.frame inherits from list)
}
}
