% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_false_positive_rate.R
\name{batch_false_positive_rate}
\alias{batch_false_positive_rate}
\title{Compute the false positive rate for a batch of synthetic experiments}
\usage{
batch_false_positive_rate(batch_output, aggregated = TRUE,
  count_attr = "count")
}
\arguments{
\item{batch_output}{A list of lists of executed \code{synthetic_experiment} objects, as returned
by the \code{batch_experiment} function.}

\item{aggregated}{A logical flag. Defaults to \code{TRUE}.}

\item{count_attr}{The name of the "count" attribute used to store the number of terms in the
computation of the average false positive rate. The default value is
\code{count}.}
}
\value{
The average false positive rate for the batch (with a count attribute
containing the number of individual resamplings of the data involved) or, if
aggregated is \code{FALSE}, a list with an average false positive rate for
each dataset (each with a count attribute).
}
\description{
Compute the false positive rate for a batch of synthetic experiments
}
\examples{
\dontrun{
batch_output <- batch_experiment("mtcars",
                                 corruptions = sample_patch_identity,
                                 N = 2,
                                 data_reader = get)
batch_false_positive_rate(batch_output)
}
}
\seealso{
\code{\link{batch_experiment}}
}
