% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_summary.R
\name{batch_summary}
\alias{batch_summary}
\title{Summarise the results of a batch of synthetic experiments}
\usage{
batch_summary(batch_output, metrics, aggregated = TRUE,
  count_attr = "count")
}
\arguments{
\item{batch_output}{A list of lists of executed \code{synthetic_experiment} objects, as returned
by the \code{batch_experiment} function.}

\item{metrics}{(Optional) A list of metric calculation functions. If missing, all metrics
will be included.}

\item{aggregated}{A logical flag. If \code{TRUE} the return value is a matrix with metrics
down the rows and corruption types across the columns. Otherwise an array is
returned containing additional dimensions for the dataset and index of the
realised corruption. Defaults to \code{TRUE}.}

\item{count_attr}{The name of the "count" attribute used to store the number of terms in the
computation of the certain metrics (not including \code{NA}s). The default value is
\code{count}.}
}
\value{
A data frame summarising the results of the batch experiment or, if
aggregated is \code{FALSE}, a list with two array elements containing,
respectively, the disaggregated metric and count results (from which the
aggregated results are calculated).
}
\description{
Summarise the results of a batch of synthetic experiments
}
\seealso{
\code{\link{batch_experiment}}
}
