% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broadband2013}
\alias{broadband2013}
\title{UK home broadband performance 2013.}
\format{A data frame with 2390 rows and 29 variables.}
\source{
\url{https://data.gov.uk/dataset/uk-fixed-line-broadband-performance}
}
\usage{
broadband2013
}
\description{
UK home broadband performance data for 2013 published by Ofcom.
}
\keyword{datasets}
