% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broadband2014}
\alias{broadband2014}
\title{UK home broadband performance 2014.}
\format{A data frame with 1971 rows and 31 variables.}
\source{
\url{https://data.gov.uk/dataset/uk-fixed-line-broadband-performance}
}
\usage{
broadband2014
}
\description{
UK home broadband performance data for 2014 published by Ofcom.
}
\keyword{datasets}
