% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{broadband2015}
\alias{broadband2015}
\title{UK home broadband performance 2015.}
\format{A data frame with 2802 rows and 66 variables.}
\source{
\url{https://data.gov.uk/dataset/uk-fixed-line-broadband-performance}
}
\usage{
broadband2015
}
\description{
UK home broadband performance data for 2015 published by Ofcom.
}
\keyword{datasets}
