% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnwise_ks.R
\name{columnwise_ks}
\alias{columnwise_ks}
\title{Perform a column-wise two-sample Kolmogorov-Smirnov test}
\usage{
columnwise_ks(df1, df2, alpha)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{alpha}{A scalar double value specifying a confidence level.}
}
\value{
A logical vector of length equal to \code{min(ncol(df1), ncol(df2))}
indicating whether the null hypothesis was rejected (\code{TRUE}) or not
(\code{FALSE}) at the given confidence level.
}
\description{
Performs a \href{https://en.wikipedia.org/wiki/Kolmogorov-Smirnov_test}{two-sample
Kolmogorov-Smirnov test} with a given confidence level \eqn{alpha}.
}
\seealso{
\code{\link{ks}}
}
