% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_columns}
\alias{compare_columns}
\title{Compare column metadata between two data frames}
\usage{
compare_columns(x, y)
}
\arguments{
\item{x, y}{Data frames to compare.}
}
\value{
A data frame of column metadata differences between
\code{x} and \code{y}.
}
\description{
Compare column metadata between two data frames
}
