% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_data}
\alias{compare_data}
\title{Compare two data frames}
\usage{
compare_data(
  x,
  y,
  context_rows = c(3L, 3L),
  context_cols = everything(),
  max_differences = Inf
)
}
\arguments{
\item{x, y}{Data frames to compare}

\item{context_rows}{Integer vector of length two indicating the number of context
row to include before and after a difference row.}

\item{context_cols}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns to include as context.}

\item{max_differences}{Maximum number of differences to return.}
}
\value{
Data frame of \code{x}, \code{y} full joined by row number. Shared variables
have suffixes as specified by \code{suffix}. A \code{.row} helper variable indicates
the row number. A \code{.type} helper variable indicates whether the row is in
\verb{x only}, \verb{y only} or \code{both} data frames.

Data frame of observations that are different in \code{x} and \code{y}, or
observations that are in only \code{x} or \code{y}, along with context rows.
}
\description{
Compare two data frames
}
