% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_join}
\alias{compare_join}
\title{Join two data frames for comparison}
\usage{
compare_join(x, y, suffix = c(".x", ".y"))
}
\arguments{
\item{x, y}{Data frames to compare}

\item{suffix}{Two element character vector of suffixes to disambiguate columns
from \code{x}, \code{y}.}
}
\value{
Data frame of \code{x}, \code{y} full joined by row number. Shared variables
have suffixes as specified by \code{suffix}. A \code{.row} helper variable indicates
the row number. A \code{.type} helper variable indicates whether the row is in
\verb{x only}, \verb{y only} or \code{both} data frames.
}
\description{
Join two data frames for comparison
}
