% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_synthetic_experiment.R
\name{configure_synthetic_experiment}
\alias{configure_synthetic_experiment}
\title{Configure a synthetic experiment}
\usage{
configure_synthetic_experiment(data_id, corruption = sample_patch_identity,
  datadiff = ddiff, N = 20, split = 0.5,
  seed = sample.int(.Machine$integer.max, size = 1), data_reader = get)
}
\arguments{
\item{data_id}{A data frame or a character vector identifying a data frame. In the latter
case, the \code{read_data} argument must contain a function which returns a
data frame when passed the vector of arguments in \code{data_id}.}

\item{corruption}{(Optional) A patch sampler function (or a list of such) to be used to generate
a patch with which to corrupt the second part of the random partition of the
\code{data}. If a list of sampler functions is provided then the generated
patches will be composed to obtain the corruption. Defaults to the identity
patch sampler (i.e. no corruption).}

\item{datadiff}{The datadiff function for the experiment.}

\item{N}{The number of experiments (i.e. random splits of the data). Defaults to 20.}

\item{split}{A number in the unit interval specifying the splitting ratio.}

\item{seed}{A random seed. By default an integer seed is chosen at random.}

\item{data_reader}{A function which reads the data, given the arguments in \code{data_id}.
Defaults to the \code{\link{get}} function.}
}
\value{
A \code{synthetic_experiment} object.
}
\description{
Generates a \code{synthetic_experiment} object, configured according
to the given parameters.
}
\details{
The \code{seed} argument provides reproducibility in the context of random
sampling from patch distributions to obtain corruptions and random splitting
of the data.

The \code{data_id} argument may be a data frame, the name of a data frame, or
a character vector specifying a data frame to be read via the given
\code{read_data} function, which itself defaults to \code{get()}.
If the \code{data_id} is a
character identifier the data are read lazily and may be stripped to save
space by calling \code{strip_data} on the returned \code{synthetic_experiment}
object.

To execute a synthetic experiment, pass the return value to the function
\code{\link{execute_synthetic_experiment}}.
}
\examples{
# Option 1: data frame passed as data_id argument. May bloat the config.
config <- configure_synthetic_experiment(data_id = mtcars)
head(config$get_data())

# Option 2: custom data_reader argument (to avoid config bloat)
data_id <- "mtcars"
config <- configure_synthetic_experiment(data_id, data_reader = get)
head(config$get_data())

\dontrun{
# Option 3: default data_reader argument is the \\code{read_data} function.
data_id <- c("source" = "uci", "dataset" = "iris")
config <- configure_synthetic_experiment(data_id)
head(config$get_data())

data_id <- c("source" = "data.gov.uk", "dataset" = "broadband", year = 2013)
config <- configure_synthetic_experiment(data_id)
head(config$get_data())
}
}
\seealso{
\code{\link{split_data}} \code{\link{execute_synthetic_experiment}}
}
