% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_scale.R
\name{cost_scale}
\alias{cost_scale}
\title{Compute cost scale factor}
\usage{
cost_scale(nx, ny)
}
\arguments{
\item{nx, ny}{The lengths of the vectors}
}
\value{
A number
}
\description{
The scale factor decreases the cost in proportion to the inverse square root
of the number of rows. This means that, when comparing against an unscaled
diffness, small diffnesses count more the more data points there are.
}
\details{
TODO: explain the origin of the scale factor.
}
