% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_matrix.R
\name{costs_matrix}
\alias{costs_matrix}
\title{Compute a costs matrix}
\usage{
costs_matrix(df1, df2 = df2, mismatch = diffness,
  patch_generators = list(gen_patch_transform), patch_penalties = 0.6,
  permute_penalty = 0.1, break_penalty = 0.99,
  penalty_scaling = purrr::partial(ks_scaling, nx = nrow(df1), ny =
  nrow(df2)), square = TRUE, verbose = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{mismatch}{Mismatch method. The default is \code{\link{diffness}}.}

\item{patch_generators}{A list of patch generator functions from which candidate patches will be
generated for every pair of columns (one each from \code{df1} & \code{df2}).}

\item{patch_penalties}{A numeric vector of patch penalties corresponding to the \code{patch_generators}
list. The lengths of these two arguments must be equal.}

\item{permute_penalty}{The penalty associated with a permutation patch. This penalty is scaled, using
the \code{penalty_scaling} argument, and multiplied by the number of columns
whose index is changed by the permutation.}

\item{break_penalty}{The penalty associated with a break patch.}

\item{penalty_scaling}{A function to be used to scale the penalty associated with each patch.
Defaults to \code{\link{ks_scaling}}.}

\item{square}{A logical flag. If \code{TRUE}, and \code{df1} & \code{df2} have differing
numbers of columns, rows will be inserted or deleted in the cost matrix,
corresponding to insertions/deletions of columns in \code{df1} (see
\code{\link{ddiff}}), to obtain a square matrix.}

\item{verbose}{A logical flag.}
}
\value{
A numeric matrix.
}
\description{
Generates a candidate patch for each pair of columns (one each from
\code{df1} & \code{df2}), calculates the corresponding costs and returns them
in a matrix.

This function is included solely for the purpose of assisting demonstration
of the datadiff package and contains code which is repeated elsewhere.
}
