% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle.R
\name{cycle}
\alias{cycle}
\title{Permute a vector according to a cycle}
\usage{
cycle(v, cyc)
}
\arguments{
\item{v}{A vector}

\item{cyc}{A cycle.}
}
\value{
The vector \code{v} with its elements permuted according to the cycle \code{cyc}.
}
\description{
Returns the vector \code{v} with its elements permuted according to the given cycle. A cycle of, e.g., c(1, 3, 4) is interpreted as the function \eqn{\sigma} such that \eqn{\sigma(1) = 3}, \eqn{\sigma(3) = 4}, \eqn{\sigma(4) = 1} and \eqn{\sigma(x) = x} for all other \eqn{x}.
}
\examples{
cycle(1:5, cyc = c(1, 3, 4))

}
