% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff.R
\name{ddiff}
\alias{ddiff}
\title{Diff two data frames}
\usage{
ddiff(df1, df2, mismatch = diffness,
  patch_generators = list(gen_patch_transform), patch_penalties = 0.6,
  permute_penalty = 0.1, break_penalty = 0.9999,
  penalty_scaling = purrr::partial(ks_scaling, nx = nrow(df1), ny =
  nrow(df2)), insert_col_name = "INSERT", as.list = FALSE,
  verbose = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{mismatch}{Mismatch method. The default is \code{\link{diffness}}.}

\item{patch_generators}{A list of patch generator functions from which, for each pair of columns (one
each from \code{df1} & \code{df2}), candidate patches will be generated.}

\item{patch_penalties}{A numeric vector of patch penalties corresponding to the \code{patch_generators}
list. The lengths of these two arguments must be equal.}

\item{permute_penalty}{The penalty associated with a permutation patch. This penalty is scaled, using
the \code{penalty_scaling} argument, and multiplied by the number of columns
whose index is changed by the permutation.}

\item{break_penalty}{The penalty associated with a break patch.}

\item{penalty_scaling}{A function to be used to scale the penalty associated with each patch.
Defaults to \code{\link{ks_scaling}}.}

\item{insert_col_name}{The column name used for inserted columns, if \code{df1} has fewer columns
than \code{df2}. Defaults to 'INSERT'.}

\item{as.list}{A logical flag. If \code{TRUE} the return value is a list of patches.
Otherwise a composite patch object is returned.}

\item{verbose}{A logical flag.}
}
\value{
A composite patch object or, if \code{as.list} is \code{TRUE}, its
decomposition as a list of elementary patches.
}
\description{
Given two data frames and a collection of patch generator functions, with
associated (numeric) penalties, this function attempts to identify a
sequence of transformations to best explain the difference and returns the
result as a (composite) patch.
}
