% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdata.R
\name{diffmeta}
\alias{diffmeta}
\title{Diff Data Frames Metadata}
\usage{
diffmeta(x, y)
}
\arguments{
\item{x, y}{Data frames to diff.}
}
\value{
A data frame of column metadata differences between
\code{x} and \code{y}.
}
\description{
Diff Data Frames Metadata
}
