% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.character}
\alias{diffness.character}
\title{Compute the mismatch between two character vectors.}
\usage{
\method{diffness}{character}(x, y, diff = tv, ...)
}
\arguments{
\item{x, y}{A pair of vectors of type \code{character}.}

\item{diff}{Mismatch method. The default is \code{tv} (total variation distance) for
unordered categorical data.}

\item{...}{Additional arguments are ignored.}
}
\description{
Character vectors are treated as unordered categorical data.
}
