% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.data.frame}
\alias{diffness.data.frame}
\title{Compute the mismatch between two data frames}
\usage{
\method{diffness}{data.frame}(x, y, col_diff = 1, ...)
}
\arguments{
\item{x, y}{A pair of data frames.}

\item{col_diff}{A numeric scalar specifying the additional mismatch per column when \code{x}
and \code{y} have different numbers of columns. Defaults to 1.}

\item{...}{Additional arguments passed to other \code{diffness} methods.}
}
\description{
Compute the mismatch between two data frames
}
