% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.double}
\alias{diffness.double}
\title{Compute the mismatch between two continuous numeric vectors}
\usage{
\method{diffness}{double}(x, y, scale = TRUE, diff = ks, ...)
}
\arguments{
\item{x, y}{A pair of vectors of type \code{double}.}

\item{scale}{A logical flag. If \code{TRUE} (the default) the result is scaled by the
length of the data.}

\item{diff}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for continuous
numeric data.}

\item{...}{Additional arguments are ignored.}
}
\description{
Compute the mismatch between two continuous numeric vectors
}
