% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.integer}
\alias{diffness.integer}
\title{Compute the mismatch between two integer vectors}
\usage{
\method{diffness}{integer}(x, y, diff = ks, ...)
}
\arguments{
\item{x, y}{A pair of vectors of type \code{integer}.}

\item{diff}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for integer
data.}

\item{...}{Additional arguments are ignored.}
}
\description{
Integer vectors are treated as ordered categorical data.
}
