% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.ordered}
\alias{diffness.ordered}
\title{Compute the mismatch between two vectors of ordered categorical data}
\usage{
\method{diffness}{ordered}(x, y, diff = ks, ...)
}
\arguments{
\item{x, y}{A pair of ordered factors.}

\item{diff}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for ordered
categorical data.}

\item{...}{Additional arguments are ignored.}
}
\description{
Compute the mismatch between two vectors of ordered categorical data
}
