% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness}
\alias{diffness}
\title{Compute the mismatch between two datasets}
\usage{
diffness(x, y, scale, ...)
}
\arguments{
\item{x, y}{A pair of datasets}

\item{scale}{A logical flag. If \code{TRUE} (the default) the result is scaled by the
size of the data.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A non-negative number quantifying the mismatch between the two
datasets.
}
\description{
\code{diffness} is a generic function: the type of the first argument
determines which method is called.
}
