% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness_scale}
\alias{diffness_scale}
\title{Compute "mismatch scale factor"}
\usage{
diffness_scale(nx, ny)
}
\arguments{
\item{nx, ny}{The lengths of the vectors}
}
\value{
A number
}
\description{
The scale factor increases the diffness in proportion to the square root of
the number of rows. This means that, when comparing against a "patch cost",
small diffnesses count more the more data points there are.
}
