% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_synthetic_experiment.R
\name{execute_synthetic_experiment}
\alias{execute_synthetic_experiment}
\title{Execute a synthetic experiment}
\usage{
execute_synthetic_experiment(config,
  logfile = system.file("log/datadiff.log", package = "datadiff"))
}
\arguments{
\item{config}{A \code{synthetic_experiment} object specifying the experiment configuration.}

\item{logfile}{The full path to a logfile. Defaults to \code{log/datadiff.log} under the
package installation directory.}
}
\value{
An executed \code{synthetic_experiment} object.
}
\description{
Executes a synthetic experiment based on a given configuration object and
returns the same object with results attached, accessible via the
\code{get_results} method.
}
\examples{

\dontrun{
corruption <- list(sample_patch_identity, sample_patch_permute)
config <- configure_synthetic_experiment(mtcars, corruption = corruption, N = 2)
execute_synthetic_experiment(config)
}
}
\seealso{
\code{\link{configure_synthetic_experiment}}
}
