% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_experiment.R
\name{executed}
\alias{executed}
\title{Determine whether a synthetic experiment has been executed}
\usage{
executed(x)
}
\arguments{
\item{x}{A \code{synthetic_experiment} object}
}
\description{
Determine whether a synthetic experiment has been executed
}
\examples{
config <- configure_synthetic_experiment("mtcars", N = 2, data_reader = get)
executed(config)

\dontrun{
execute_synthetic_experiment(config)
executed(config)
}

}
