% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_canonical_permutation.R
\name{extract_canonical_permutation}
\alias{extract_canonical_permutation}
\title{Extract the canonical elementary permutation from a patch}
\usage{
extract_canonical_permutation(patch, perm_param = "perm",
  param_only = FALSE)
}
\arguments{
\item{patch}{A patch.}

\item{perm_param}{The name of the permutation parameter in patches of the relevant types.}

\item{param_only}{A logical flag. If \code{TRUE}, the \code{perm} parameter value is returned,
rather than a permuation patch having that parameter. Defaults to \code{FALSE}.}
}
\value{
An elementary permutation patch.
}
\description{
The canonical elementary permutation is the unique permutation such that
applying the following procedure:
 1. form the patch \code{p} which is identical to the given \code{patch}
    except any permutation element is removed,
 2. apply the patch \code{p},
 3. apply the canonical elementary permutation,
produces the same column structure (in the resulting data frame) as that
produced by applying the given \code{patch}.
}
\details{
Note that it is assumed that only the insert and delete patches modify
the number of columns in the data frame and only the permutation patch
modifies the column order (i.e. there are no other patch types that modify
the data frame column structure).
}
\examples{
patch <- compose_patch(patch_delete(cols = 4L), patch_permute(perm = c(2:11, 1L)))
canonical_patch <- extract_canonical_permutation(patch)
colnames(patch(mtcars))
# To construct the equivalent patch we must delete the column that's
# in position 4 after the original permutation, i.e. column 5.
equiv_patch <- compose_patch(canonical_patch, patch_delete(cols = 5L))
colnames(equiv_patch(mtcars))
}
