% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_batch_summary.R
\name{format_batch_summary}
\alias{format_batch_summary}
\title{Format a summary of batched synthetic experiments}
\usage{
format_batch_summary(batch_summary, filename, path, tex = FALSE)
}
\arguments{
\item{batch_summary}{A batch output summary matrix of the type returned by the
\code{\link{batch_summary}} function.}

\item{filename}{A file name.}

\item{path}{A file path.}

\item{tex}{A logical flag. If \code{TRUE} the table rows are formatted for TeX. Defaults
to \code{FALSE}.}
}
\description{
Writes a markdown formatted table containing the summarised batch experiment
results to a text file.
}
\seealso{
\code{\link{batch_experiment}}
}
\keyword{internal}
