% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_affine.R
\name{gen_patch_affine}
\alias{gen_patch_affine}
\title{Generate an affine patch}
\usage{
gen_patch_affine(df1, df2, mismatch = ks, col1, col2 = col1,
  verbose = FALSE, ...)
}
\arguments{
\item{df1}{A data frame. The column specified in the \code{col1} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{df2}{A data frame. The column specified in the \code{col2} argument must contain
a vector of type \code{double} with at least one two-missing values.}

\item{mismatch}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov).}

\item{col1}{A column identifier (integer or string column name) of length 1.}

\item{col2}{A column identifier (integer or string column name) of length 1. By default
this takes the value of \code{col1}.}

\item{verbose}{A logical flag to turn on/off warnings associated with failed optimisation
attempts.}

\item{...}{Additional arguments are ignored.}
}
\value{
An 'affine' \code{patch} object (consisting of the composition of a
shift and a scale patch).
}
\description{
Generates a composite \code{patch} object representing an affine
transformation, formed as the composition of a \code{patch_shift} and a
\code{patch_scale} object, whose \code{shift} and \code{scale_factor} parameters
have been selected with the aim of minimising the mismatch between the specified
columns after application of the patch to \code{df1}.
}
\details{
Uses the \code{\link{optim}} function to optimise the affine transformation
parameters for the given mismatch method. The numerical methods available to
\code{optim} are tried, in the order in which they apper in that function's
default \code{method} argument, until convergence is achieved. If none of the
methods succeeds an error is thrown.
}
\examples{
gen_patch_affine(mtcars, mtcars, col1 = "wt", col2 = "qsec")

}
\seealso{
\code{\link{optim}}
}
