% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_break.R
\name{gen_patch_break}
\alias{gen_patch_break}
\title{Generate a break patch}
\usage{
gen_patch_break(df1, df2, col1, col2 = col1, ...)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame.}

\item{col1}{A column identifier (integer or string column name) of length 1.}

\item{col2}{A column identifier (integer or string column name) of length 1. By default
this takes the value of \code{col1}.}

\item{...}{Additional arguments are ignored.}
}
\value{
A \code{patch_break} object.
}
\description{
Generates a \code{patch_break} object.
}
\seealso{
\code{\link{patch_break}}
}
