% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_identity.R
\name{gen_patch_identity}
\alias{gen_patch_identity}
\title{Generate an identity patch}
\usage{
gen_patch_identity(df1, df2, mismatch = purrr::partial(diffness, scale =
  FALSE), ...)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame.}

\item{mismatch}{Mismatch method. The default is (unscaled) \code{\link{diffness}}.}

\item{...}{Additional arguments are ignored.}
}
\value{
An identity \code{patch} object with mismatch method attached as an
attribute.
}
\description{
Generates an identity patch object with mismatch method attached as an
attribute.
}
\seealso{
\code{\link{optim}}
}
