% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_insert.R
\name{gen_patch_insert}
\alias{gen_patch_insert}
\title{Generate an insert patch}
\usage{
gen_patch_insert(df1, df2, col1 = as.integer(col2 - 1), col2,
  prefix = "INSERT.", ...)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame.}

\item{col1}{A column identifier (integer or string column name) of length 1, specifying
the \code{insertion_point} of the new column in \code{df1}. By default this
takes the value of \eqn{col2 - 1}, so the column \emph{index} of the newly
inserted column in \code{df1} is the same as that of the \code{col2} in
\code{df2}.}

\item{col2}{A column identifier (integer or string column name) of length 1.}

\item{prefix}{A character string specifying the prefix to be added to the inserted column
name. The rest of the name is equal to that of \code{col2} in \code{df2}.}

\item{...}{Additional arguments are ignored.}
}
\value{
A \code{patch_insert} object.
}
\description{
Generates a \code{patch_insert} object which, when applied to \code{df1},
inserts a new column into that data frame (containing only \code{NA} values)
whose name matches that of \code{col2} in \code{df2}, prefixed by the string
given in the \code{prefix} argument.
}
\seealso{
\code{\link{patch_insert}}
}
