% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_perm.R
\name{gen_patch_perm}
\alias{gen_patch_perm}
\title{Generate an optimal permutation patch}
\usage{
gen_patch_perm(df1, df2, cost = 0)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{A data frame}

\item{cost}{A "permutation cost"}
}
\value{
A \code{patch_perm} object.
}
\description{
Computes the permutation of columns, expressed as a \code{patch_perm} patch,
that minimises the mismatch between the patched \code{df1} and \code{df2},
including a permutation cost. Currently the permutation cost is ignored.
}
\details{
This function uses an assignment problem solver from the package
\code{lpSolve}.
}
