% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_recode.R
\name{gen_patch_recode}
\alias{gen_patch_recode}
\title{Generate a recode patch}
\usage{
gen_patch_recode(df1, col1, df2, col2 = col1, ...)
}
\arguments{
\item{df1}{A data frame.}

\item{col1}{A column identifier (integer or string column name) with length 1.}

\item{df2}{A data frame.}

\item{col2}{A column identifier (integer or string column name) with length 1. By default
this takes the value of \code{col1}.}

\item{...}{Additional arguments passed to the \code{patch_recode} function.}
}
\value{
A \code{patch_recode} object.
}
\description{
Generates a \code{patch_recode} object whose 'encoding' parameter has been
selected with the aim of minimising the mismatch between the specified
columns after application of the patch to \code{df1}.
}
\seealso{
\code{\link{patch_recode}}
}
