% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_shift.R
\name{gen_patch_shift}
\alias{gen_patch_shift}
\title{Generate a shift patch}
\usage{
gen_patch_shift(df, cols, y)
}
\arguments{
\item{df}{A data frame. The column specified in the \code{cols} argument must contain a
vector of type \code{double} with at least one non-missing value.}

\item{cols}{A column identifier (integer or string column name) with length 1.}

\item{y}{A 'target' vector of type \code{double} with least one non-missing value.}

\item{...}{Additional arguments passed to the \code{mean} function.}
}
\value{
A \code{patch_shift} object.
}
\description{
Generates a \code{patch_shift} object whose 'shift' parameter has been
selected with the aim of minimising the mismatch between the vector \code{y}
and the patch applied to the given data frame.
}
\seealso{
\code{\link{gen_patch_affine}}
}
