% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_transform.R
\name{gen_patch_transform}
\alias{gen_patch_transform}
\title{Generate a transformation patch}
\usage{
gen_patch_transform(df1, df2, mismatch = diffness, col1, col2 = col1,
  verbose = FALSE, ...)
}
\arguments{
\item{df1}{A data frame. The column specified in the \code{col1} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{df2}{A data frame. The column specified in the \code{col2} argument must contain
a vector of type \code{double} with at least one two-missing values.}

\item{mismatch}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for continuous
data and \code{tv} (Total variation distance) for categorical data.}

\item{col1}{A column identifier (integer or string column name) of length 1.}

\item{col2}{A column identifier (integer or string column name) of length 1. By default
this takes the value of \code{col1}.}

\item{verbose}{A logical flag to turn on/off warnings associated with failed optimisation
attempts in the continuous case.}

\item{...}{Additional arguments are ignored.}
}
\value{
A \code{patch} object, or \code{NULL} in case of error.
}
\description{
Generates either an 'affine' \code{patch} object, in case of continuous data,
or a 'recode' \code{patch} object in the case of categorical data.
}
\seealso{
\code{\link{gen_patch_affine}} \code{\link{gen_patch_recode}}
}
