% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator_type.R
\name{generator_type}
\alias{generator_type}
\title{Extract the patch type from a columnwise patch generator function}
\usage{
generator_type(df, patch_generator, short = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{patch_generator}{A patch generator function or a list of such.}

\item{short}{A logical flag passed to the \code{\link{patch_type}} function. Defaults to
\code{TRUE}.}
}
\value{
A character vector containing the patch type corresponding to the
\code{patch_generator} (whose order corresponds to that of the
\code{patch_generator} in case it is a list of generator functions). An
\code{NA} value indicates that the corresponding generator function is
incompatible with \emph{all} columns in the data, and therefore the patch
type could not be deduced.
}
\description{
Deduces the patch type of an elementary patch generator function by applying
the generator to the given data (after splitting, using \code{split_data}).
}
\details{
Vectorised over the \code{patch_generator} argument (which may be a list).
}
\examples{
generator_type(mtcars, gen_patch_rescale)
generator_type(esoph, list(gen_patch_rescale, gen_patch_recode))

}
