% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{get_patch_params}
\alias{get_patch_params}
\title{Get the parameters associated with a patch object.}
\usage{
get_patch_params(patch)
}
\arguments{
\item{patch}{A \code{patch} object.}
}
\value{
A list containing the parameters associated with the given
\code{patch}. If \code{patch} is a composition of patches then a list of
lists is returned.
}
\description{
Returns the \code{patch} parameters object.
}
\examples{
get_patch_params(patch_shift(2L, shift = 10))
get_patch_params(patch_rescale(1L, shift = 0, scale_factor = 2))
get_patch_params(patch_identity())

}
