% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heartdisease}
\alias{heartdisease}
\title{data from the UCI Machine Learning Repository}
\format{A data frame with 303 rows and 14 variables.}
\source{
\url{http://archive.ics.uci.edu/ml}
}
\usage{
heartdisease
}
\description{
UK home broadband performance data for 2015 published by Ofcom.
}
\keyword{datasets}
