% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Iris data set}
\format{A data frame with 4177 rows and 9 variables.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Iris}
}
\usage{
iris
}
\description{
Iris data set from the UCI Machine Learning Repository.
}
\keyword{datasets}
