% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_columnwise_unique.R
\name{is_columnwise_unique}
\alias{is_columnwise_unique}
\title{Test for columnwise uniqueness}
\usage{
is_columnwise_unique(patch, type = NULL, short = TRUE,
  column_param = "cols")
}
\arguments{
\item{patch}{A patch object.}

\item{type}{A patch type or vector of patch types. If \code{NULL} (the default), all
relevant patch types are tested.}

\item{short}{A logical flag passed to the \code{\link{patch_type}} function. If
\code{TRUE} (the default) the \code{type} argument must specify the patch
type in short form.}

\item{column_param}{The name of the column parameter in patches of the specified type.}
}
\value{
A logical vector of length equal to the \code{type} argument (or, if
\code{type} is \code{NULL}, the number of relevant elementary types in the
given \code{patch}). Each element is \code{TRUE} if the given patch is
columnwise unique for the corresponding type; otherwise \code{FALSE}. If the
\code{type} argument has length greater than one, the elements of the return
value are named by the corresponding type.
}
\description{
A composed patch is \emph{columnwise unique} for a particular patch type
if its decomposition into elementary patches contains at most one patch of
the given type applied to any particular column. An error is thrown if the
given \code{patch} does not contain any elementary component of the given
\code{type}.
}
\details{
The \code{\link{terminal_column_position}} function is used to uniquely
identify each column because column indices are subject to modification
in the presence of insert, delete or permutation patches.
}
\examples{
patch <- compose_patch(patch_rescale(3L, shift = 0, scale_factor = 2),
                       patch_delete(3L),
                       patch_rescale(4L, shift = 1, scale_factor = 10))
# Two rescale patches are applied to the same column:
is_columnwise_unique(patch)

patch <- compose_patch(patch_recode(3L, encoding = c("N" = 0, "Y" = 1)),
                       patch_delete(3L),
                       patch_rescale(4L, shift = 1, scale_factor = 10))
# Two patches of different types are applied to the same column:
is_columnwise_unique(patch)

}
\seealso{
\code{\link{terminal_column_position}}
}
