% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.patch.R
\name{is_compatible.delete.patch}
\alias{is_compatible.delete.patch}
\title{Test for compatibility between a patch and a data frame.}
\usage{
\method{is_compatible}{delete.patch}(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Any additional arguments are ignored.}
}
\description{
Implements the generic function \code{is_compatible} for a \code{patch} of
type \code{delete.patch}. Returns \code{TRUE} if the given patch and data
frame are compatible, in the sense that the patch (function) may be applied
to the data frame without generating an error.

The patch is deemed compatible if the maximum integer (in absolute value) in
its \code{params} object does not exceed the number of columns in the data
frame.
}
\examples{
ncol(mtcars)
p <- patch(-2L)
is_compatible(p, mtcars)
p <- patch(-12L)
is_compatible(p, mtcars)

}
