% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is_compatible}
\alias{is_compatible}
\title{Test for compatibility between patch and data frame}
\usage{
is_compatible(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Additional arguments.}
}
\description{
Test for compatibility between patch and data frame
}
