% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.patch.R
\name{is_compatible.scale.patch}
\alias{is_compatible.scale.patch}
\title{Test for compatibility between patch and data frame.}
\usage{
\method{is_compatible}{scale.patch}(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Any additional arguments are ignored.}
}
\description{
Implements the generic function \code{is_compatible} for a \code{patch} of
type \code{scale.patch}. Returns \code{TRUE} if the given patch and data
frame are compatible, in the sense that the patch (function) may be applied
to the data frame without generating an error.

The \code{patch} is deemed compatible if both:
\itemize{
 \item the maximum integer in the \code{datadiff:::COLUMNS} element of its
 \code{params} object does not exceed the number of columns in the data
 frame, and
 \item all of the columns in the data frame whose indices are specified in
 the \code{datadiff:::COLUMNS} element contain numerical data.
}
}
\examples{
# Columns 1 and 3 in the ToothGrowth data frame are numeric.
p <- patch(scale_patch_params(c(1L, 3L), 2))
is_compatible(p, ToothGrowth)
# Column 2 is not.
p <- patch(scale_patch_params(c(1L, 2L), 2))
is_compatible(p, ToothGrowth)

}
