% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_compatible_columns.R
\name{is_compatible_columns}
\alias{is_compatible_columns}
\title{Test for compatibility between a data frame and column identifiers}
\usage{
is_compatible_columns(cols, df)
}
\arguments{
\item{cols}{A vector of column identifiers (character or integer).}

\item{df}{A data frame.}
}
\value{
\code{TRUE} if the given column identifiers are compatible with the given
data frame, otherwise \code{FALSE}.
}
\description{
Returns \code{TRUE} if the given column identifiers are compatible with the
given data frame.
}
\examples{
is_compatible_columns(c(2L, 5L), mtcars)
is_compatible_columns(c("mpg", "gear"), mtcars)

is_compatible_columns(c(2L, 22L), mtcars)
is_compatible_columns(c("mpg", "xxx"), mtcars)

}
