% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{is_equal}
\alias{is_equal}
\title{Equality Test}
\usage{
is_equal(x, y, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{Vectors to compare.}

\item{tol}{Numeric tolerance for comparison.}
}
\value{
TRUE if equal (within tolerance), FALSE otherwise.
}
\description{
Equality Test
}
