% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is_identity_patch}
\alias{is_identity_patch}
\title{Test for an identity patch object}
\usage{
is_identity_patch(obj, allow_composed = TRUE)
}
\arguments{
\item{obj}{Any object.}

\item{allow_composed}{A logicial flag. If \code{TRUE} (the default) then compositions of identity
patches (constructed using either the \code{compose_patch} function or the
\code{compose} function from the \code{purrr} package) are considered to be
identity patches. Otherwise, only an elementary identity patch object is
considered to be an identity patch.}
}
\description{
Returns \code{TRUE} if the \code{obj} argument is the identity patch or, if
\code{allow_composed} is \code{TRUE}, a composition of identity patches.
}
\examples{
is_identity_patch(patch_identity())
is_identity_patch(compose_patch(patch_identity(), patch_identity()))
is_identity_patch(compose_patch(patch_identity(), patch_identity()),
                  allow_composed = FALSE)
# Patches of other types are not considered to be the identity,
# regardless of their parameter values.
is_identity_patch(patch_rescale(1L, shift = 0, scale_factor = 1))
is_identity_patch(patch_permute(1:10))
is_identity_patch(mtcars)

}
