% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is_patch}
\alias{is_patch}
\title{Test for a patch object}
\usage{
is_patch(obj, allow_composed = TRUE)
}
\arguments{
\item{obj}{Any object.}

\item{allow_composed}{A logicial flag. If \code{TRUE} (the default) then compositions of patches
constructed using the \code{purrr::compose} function are considered to be
patches. Otherwise, only elementary patch objects are considered to be
patches.}
}
\description{
Returns \code{TRUE} if the \code{obj} argument is a \code{patch} object.
}
