% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_experiment.R
\name{is_synthetic_experiment}
\alias{is_synthetic_experiment}
\title{Test for a synthetic experiment object}
\usage{
is_synthetic_experiment(obj)
}
\arguments{
\item{obj}{Any object.}
}
\description{
Returns \code{TRUE} iff the \code{obj} argument is a \code{synthetic
experiment} object.
}
\examples{
config <- configure_synthetic_experiment("mtcars", data_reader = get)
is_synthetic_experiment(config)
is_synthetic_experiment(mtcars)

}
