% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_columns.R
\name{is_valid_columns}
\alias{is_valid_columns}
\title{Test for valid column identifiers}
\usage{
is_valid_columns(cols)
}
\arguments{
\item{cols}{Any object.}
}
\value{
\code{TRUE} if the \code{cols} object is a valid vector of column
identifiers and \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if \code{cols} is a non-empty vector of unique, non-NA
values each of which is either:
\itemize{
 \item a positive integer, or
 \item a non-empty character string
}
}
\examples{
is_valid_columns(c(2L, 5L))
is_valid_columns(c("mpg", "gear"))

# Elements must be either non-empty character strings or positive integers.
is_valid_columns(c(2, 5))
is_valid_columns(c(0L, 5L))
is_valid_columns(c("mpg", ""))

}
