% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks.numeric}
\alias{ks.numeric}
\title{K-S statistic for two numeric vectors}
\usage{
\method{ks}{numeric}(v1, v2)
}
\arguments{
\item{v1, v2}{A pair of numeric vectors. Both arguments must have at
least one non-missing value.}
}
\value{
A number between 0 and 1 inclusive.
}
\description{
Computes the Kolmogorov-Smirnov statistic for two numeric vectors by
comparing their empirical cumulative distribution functions obtained from
the \code{\link{ecdf}} function.
}
\seealso{
\code{\link{ecdf}}
}
\keyword{internal}
