% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks.ordered}
\alias{ks.ordered}
\title{K-S statistic for two ordered factors.}
\usage{
\method{ks}{ordered}(v1, v2)
}
\arguments{
\item{v1, v2}{A pair of ordered factors. Both arguments must have at least one non-missing
value. The orderings must not conflict and must together determine a unique
ordering on the union of the levels.}
}
\value{
A number between 0 and 1 inclusive.
}
\description{
Computes the Kolmogorov-Smirnov statistic for two ordered factors by
first determining a unique ordering on the union of the factor levels, if
one exists, and then comparing the empirical cumulative distribution
functions obtained from the \code{\link{ecdf}} function.
}
\seealso{
\code{\link{ecdf}}
}
