% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks}
\alias{ks}
\title{K-S statistic}
\usage{
ks(v1, v2)
}
\arguments{
\item{v1, v2}{A pair of numeric vectors or ordered factors.}
}
\value{
A number between 0 and 1 inclusive.
}
\description{
Computes the two-sample Kolmogorov-Smirnov statistic.
}
\examples{
ks(rnorm(10), rnorm(20))
ks(as.ordered(1:10), as.ordered(20:1))

}
