% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_ddiff.R
\name{ks_ddiff}
\alias{ks_ddiff}
\title{Diff two data frames}
\usage{
ks_ddiff(df1, df2, alpha = 0.1, verbose = FALSE, ...)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{alpha}{A scalar double value specifying a confidence level.}

\item{verbose}{A logical flag.}

\item{...}{Additional arguments passed to the \code{\link{ddiff}} function.}
}
\value{
A composite patch object or, if \code{as.list} is \code{TRUE}, its
decomposition as a list of elementary patches.
}
\description{
Diff two data frames
}
\examples{
\dontrun{
ks_ddiff(mtcars, mtcars[, 11:1])
}

}
