% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks_ecdf}
\alias{ks_ecdf}
\title{K-S statistic for two ecdf's}
\usage{
ks_ecdf(e1, e2)
}
\arguments{
\item{e1, e2}{A pair of ecdf objects.}
}
\value{
A number between 0 and 1 inclusive.
}
\description{
Internal function to calculate the K-S statistic for two empirical
cumulative distribution functions.
}
\keyword{internal}
