% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_scaling.R
\name{linear_scaling}
\alias{linear_scaling}
\title{Scale a penalty linearly}
\usage{
linear_scaling(penalty, nx, ny)
}
\arguments{
\item{penalty}{A numeric penalty}

\item{nx, ny}{The lengths of the vectors}
}
\value{
A (numeric) linearly-scaled penalty.
}
\description{
The scale factor is the square root of the sum of the lengths divided by the
product of the lengths.
}
\examples{
linear_scaling(0.5, nx = 100, ny = 100)

}
\seealso{
\code{\link{ks_scaling}}
}
