% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_types.R
\name{matching_types}
\alias{matching_types}
\title{Identify matching patch types}
\usage{
matching_types(patch, param_name, short = TRUE)
}
\arguments{
\item{patch}{A patch.}

\item{param_name}{A scalar parameter name.}

\item{short}{A logical flag passed to the \code{patch_type} function.}
}
\value{
The matching patch types as a character vector.
}
\description{
Returns the vector of elementary patch types found in the given \code{patch}
which have a parameter with the given name.
}
\examples{
patch <- compose_patch(patch_permute(10:1),
                       patch_rescale(1L, shift = 0, scale_factor = 2),
                       patch_delete(1L))
matching_types(patch, param_name = "shift")

}
