% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_column_accuracy.R
\name{metric_column_accuracy}
\alias{metric_column_accuracy}
\title{Column accuracy evaluation metric}
\usage{
metric_column_accuracy(experiment, partial = FALSE, column_param = c("cols",
  "insertion_point"))
}
\arguments{
\item{experiment}{An executed synthetic experiment object.}

\item{partial}{A logical flag. If \code{TRUE}, partial credit is given (see the details for
the function \code{\link{pairwise_column_accuracy}}).
Defaults to \code{FALSE}.}

\item{column_param}{A character vector containing the names of the column parameters in patches
of the relevant types. Defaults to \code{c("cols", "insertion_point")}.}
}
\value{
The observed column accuracy for each relevant corruption type as a
vector of numbers in the unit interval, with elements named by the patch type,
or \code{NA} if there are no relevant types in the corruption.
}
\description{
Returns the average column accuracy observed in a synthetic experiment.
}
\details{
This function will fail (with an error) unless the corruption in the synthetic
\code{experiment} is column-wise unique. In other words, it must
contain at most one patch of any given type applied to any given column
(see \code{\link{is_columnwise_unique}}). Although this is not strictly
required for calculating the column accuracy metric, it simplifies the
implementation and entails no loss of generality (since for any patch which
is not column-wise unique there is an equivalent one which is).
}
\examples{

\dontrun{
corruption <- sample_patch_rescale
config <- configure_synthetic_experiment(mtcars, corruption = corruption, N = 2)
output <- execute_synthetic_experiment(config)
metric_column_accuracy(output)
}
}
\seealso{
\code{\link{pairwise_column_accuracy}}
}
