% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_hamming_distance.R
\name{metric_hamming_distance}
\alias{metric_hamming_distance}
\title{Hamming distance metric for permutations}
\usage{
metric_hamming_distance(experiment, perm_param = "perm",
  count_attr = "count")
}
\arguments{
\item{experiment}{An executed synthetic experiment object.}

\item{perm_param}{The name of the permutation parameter in patches of the relevant types.}

\item{count_attr}{The name of the "count" attribute used to store the number of terms averaged
in the the computation of the mean Hamming distance. Set this argument to
\code{NULL} to omit this attribute. The default value is \code{count}.}
}
\value{
A numeric scalar containing the mean observed Hamming distance
between the permutation in the corruption and that in the results.
}
\description{
Returns the mean Hamming distance for permutations observed in a synthetic
experiment. Both the corruption and the results are assumed to contain at
most one permutation patch (otherwise an error is thrown).
}
\details{
Returns \code{NA} if there are no patch types in the elementary decomposition
of the corruption that contain the given permutation parameter.
}
\examples{

\dontrun{
corruption <- purrr::partial(sample_patch_permute, n = 2L)
config <- configure_synthetic_experiment(mtcars, corruption = corruption, N = 2)
output <- execute_synthetic_experiment(config)
metric_hamming_distance(output)
}
}
