% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_parameter_accuracy.R
\name{metric_parameter_accuracy}
\alias{metric_parameter_accuracy}
\title{Parameter accuracy metric}
\usage{
metric_parameter_accuracy(experiment, column_param = "cols",
  count_attr = "count")
}
\arguments{
\item{experiment}{An executed synthetic experiment object.}

\item{column_param}{The name of the column parameter in patches of the relevant types.}

\item{count_attr}{The name of the "count" attribute used to store the number of terms averaged
in the the computation of the mean parameter accuracy.
This attribute is intended as an aid when interpreting the results. Set this
argument to \code{NULL} to withdraw this attribute. The default value is
\code{count}.}
}
\value{
A numeric scalar containing the mean observed parameter accuracy.
}
\description{
Returns the mean parameter accuracy observed in a synthetic experiment.
Pairwise parameter accuracy is a Boolean quantity: it is \code{TRUE} iff a
pair of vector (or list) parameters define the same mapping between element
names and values. The parameter accuracy metric is the average pairwise
accuracy between the corruption and the experiment results (after converting
each Boolean pairwise result to a binary digit).
}
\details{
Returns \code{NA} if there are no patch types in the elementary decomposition
of the corruption that contain the given parameter name.
}
\seealso{
\code{\link{metric_parameter_comparison}}
}
