% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_parameter_comparison.R
\name{metric_parameter_comparison}
\alias{metric_parameter_comparison}
\title{Parameter comparison metric computation}
\usage{
metric_parameter_comparison(experiment, f_metric, f_comparison, f_map,
  f_param_match, column_param = "cols", count_attr = "count")
}
\arguments{
\item{experiment}{An executed synthetic experiment object.}

\item{f_metric}{A function for computing the comparison metric from a vector of pairwise
comparisons.}

\item{f_comparison}{A function for pairwise parameter comparison.}

\item{f_map}{A map function (e.g. \code{purrr::map_dbl} is appropriate in case
\code{f_comparison} returns a scalar double value).}

\item{f_param_match}{A predicate for matching parameters based on their value. For example, if
the parameter of interest is a double value, the \code{is.double} function
is appropriate here.}

\item{column_param}{The name of the column parameter in patches of the relevant types.}

\item{count_attr}{The name of the "count" attribute used to store the number of terms in the
computation of the metric (not including NAs). This attribute is intended as
an aid when interpreting the results. Set this argument to \code{NULL} to
withdraw this attribute. The default value is \code{count}.}
}
\value{
The observed parameter comparison metric for each relevant
patch type as a list of numeric vectors. The list elements are named according
to the patch types. The numeric vector element names correspond to the
matching parameter names (i.e. those to which the particular metric applies,
as determined by the \code{f_param_match} function).
}
\description{
Returns the parameter comparison metric observed in a synthetic experiment.
Note that the calculation of parameter comparison metrics ignores experiment
results which are not column-accurate and therefore it is important to take
into account the column accuracy metric when interpreting parameter
comparison metrics. The count attribute (see the \code{count_attr} argument)
serves a similar purpose.
}
\details{
Returns \code{NA} if there are no patch types in the elementary decomposition
of the corruption that contain matching parameters (as identified by the
\code{f_param_match} argument.
}
\seealso{
\code{\link{pairwise_parameter_comparison}}
\code{\link{metric_parameter_RMSE}} \code{\link{metric_parameter_accuracy}}
}
