% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_type_recall.R
\name{metric_type_recall}
\alias{metric_type_recall}
\title{Patch type recall evaluation metric}
\usage{
metric_type_recall(experiment, short = TRUE, count_attr = "count")
}
\arguments{
\item{experiment}{An executed synthetic experiment object.}

\item{short}{A logical flag passed to the \code{\link{patch_type}} function.}

\item{count_attr}{A character string specifying the name of the 'count' attribute.}
}
\value{
A list of numbers, each in the unit interval, each with an element
name corresponding to the patch type and a count attribute which counts the
number of terms from which the metric was calculated.
}
\description{
Returns the patch type recall rate observed in a synthetic experiment.
}
\examples{

\dontrun{
corruption <- sample_patch_rescale
config <- configure_synthetic_experiment(mtcars, corruption = corruption, N = 2)
output <- execute_synthetic_experiment(config)
metric_type_recall(output)
}
}
\seealso{
\code{\link{execute_synthetic_experiment}}
}
