% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_batch_experiment.R
\name{multi_batch_experiment}
\alias{multi_batch_experiment}
\title{Configure & execute a batch of synthetic experiments}
\usage{
multi_batch_experiment(data_ids, corruptions, datadiff = ddiff, N = 10,
  M = 10, split = 0.5, hyperseed = sample.int(.Machine$integer.max, size =
  1), data_reader = get, execute = TRUE, pb = TRUE, logfile = NULL)
}
\arguments{
\item{data_ids}{A collection of data frame identifiers. Each one must be a valid
\code{data_id} input to \code{\link{configure_synthetic_experiment}}.}

\item{corruptions}{A collection of patch sampler functions. Each one must be a valid
\code{corruption} input to \code{\link{configure_synthetic_experiment}}.}

\item{datadiff}{The datadiff function for the experiment.}

\item{N}{The number of experiments (i.e. random splits of the data) per random seed.}

\item{M}{The number of random seeds per corruption type.}

\item{split}{A number in the unit interval specifying the splitting ratio.}

\item{hyperseed}{A random seed used to select the \code{M} seeds for each corruption. By
default an integer seed is chosen at random.}

\item{data_reader}{A function which reads the data, given each of the arguments in \code{data_id}.}

\item{execute}{A logical flag. If \code{TRUE} (the default) the batch experiment will be
executed, and the outputs returned. Otherwise the batch of experiment
configurations will be returned without performing the experiments.}

\item{pb}{A logical flag. If \code{TRUE} (the default) a progress bar will be
displayed in the console.}

\item{logfile}{(Optional) The full path to a logfile.}
}
\value{
A list of lists of exectuted \code{synthetic_experiment} objects. The
outer list corresponds to the set of datasets specified in the \code{data_ids}
argument. The inner list corresponds to the list of \code{corruptions}.
}
\description{
Use this function to run synthetic experiments sequentially on multiple
datasets and with multiple corruptions. To process the results, pass the
return value to the \code{batch_summary} function.
}
\examples{
\dontrun{
data_ids <- c("mtcars", "airquality")
corruptions <- list(sample_patch_identity,
                   list(sample_patch_delete, sample_patch_permute))
multi_batch_experiment(data_ids, corruptions = corruptions, N = 2, M = 3,
                       data_reader = get)
}
}
\seealso{
\code{\link{batch_experiment}}
}
