% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{or_operator}
\alias{or_operator}
\alias{\%||\%}
\title{Operator for default values}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value to return if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns y if x is NULL, otherwise returns x
}
