% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_parameter_comparison.R
\name{pairwise_parameter_comparison}
\alias{pairwise_parameter_comparison}
\title{Compute a pairwise parameter comparison}
\usage{
pairwise_parameter_comparison(target, result, type, f_comparison,
  f_map = purrr::map_dbl, f_param_match, short = TRUE,
  column_param = "cols")
}
\arguments{
\item{target}{A patch.}

\item{result}{A patch.}

\item{type}{A patch type.}

\item{f_comparison}{A function for pairwise parameter comparison.}

\item{f_map}{A map function (e.g. \code{purrr::map_dbl} is appropriate in case
\code{f_comparison} returns a scalar double value).}

\item{f_param_match}{A predicate for matching parameters based on their value. For example, if
the parameter of interest is a double value, the \code{is.double} function
is appropriate here.}

\item{short}{A logical flag passed to the \code{\link{patch_type}} function. If
\code{TRUE} (the default) the \code{type} argument must specify the patch
type in short form.}

\item{column_param}{The name of the column parameter in patches of the specified type.}
}
\value{
A named list of non-negative vectors correpsonding to the parameters
in the given patch type. Each vector has elements whose type is determined by
\code{f_map} (or NA if the (partial) pairwise column accuracy of the
\code{result} w.r.t. the \code{target} is zero) with named elements
corresponding to the initial position of the column that is transformed by
the \code{target} and \code{result} patches. The vectors are ordered by the
element names.
}
\description{
Returns a pairwise parameter comparison of a \code{result} w.r.t. the
\code{target} patch.
}
\details{
The result is a list with a named element for each parameter associated with
the given patch \code{type}. Each list element is a vector with named elements
corresponding to the intitial positions of the columns that are transformed
by the \code{target} and \code{result} patches. Vector element values
contain the computed parameter comparison, or \code{NA} if either:
 - the initial column position appears in the target but not the result, or
 - the initial column position is moved by the result patch to a different
   position than that to which it is moved by the target patch.

Returns NA if the (partial) pairwise column accuracy of the \code{result}
w.r.t. the \code{target} is zero (i.e. if none of the columns transformed by
the target \code{patch} are correctly identified in the \code{result}), or if
the final positions of patched columns differ after application of the result
and target patches.
}
