% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_break.R
\name{patch_break}
\alias{patch_break}
\title{\code{patch_break} S3 class constructor}
\usage{
patch_break(cols)
}
\arguments{
\item{cols}{A vector of column identifiers.}
}
\value{
A \code{patch_break} object.
}
\description{
S3 class \code{patch_break} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by an abrubt break in one or more columns.
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
