% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_cycle.R
\name{patch_cycle}
\alias{patch_cycle}
\title{\code{patch_cycle} S3 class constructor}
\usage{
patch_cycle(cols)
}
\arguments{
\item{cols}{A vector of column identifiers.}
}
\value{
A \code{patch_cycle} object.
}
\description{
S3 class \code{patch_cycle} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by permutation of the column indices by a cycle.
}
\examples{
colnames(mtcars)
p <- patch_cycle(c("mpg", "gear", "cyl"))
p <- patch_cycle(c(1L, 4L))

# The following are equivalent:
colnames(apply_patch(mtcars, p))
colnames(p(mtcars))

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_cycle(c(1L, 3L, 22L))
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
\code{\link{cycle}}
}
