% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_delete.R
\name{patch_delete}
\alias{patch_delete}
\title{\code{patch_delete} S3 class constructor}
\usage{
patch_delete(cols)
}
\arguments{
\item{cols}{A vector of column identifiers.}
}
\value{
A \code{patch_delete} object.
}
\description{
S3 class \code{patch_delete} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by deletion of one or more columns.
}
\examples{
colnames(mtcars)
p <- patch_delete(c(2L, 5L))
p <- patch_delete(c("mpg", "gear"))

# The following are equivalent:
colnames(apply_patch(mtcars, p))
colnames(p(mtcars))

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_delete(22L)
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
