% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_insert.R
\name{patch_insert}
\alias{patch_insert}
\title{\code{patch_insert} S3 class constructor}
\usage{
patch_insert(insertion_point, data)
}
\arguments{
\item{insertion_point}{A scalar column identifier. The new columns will be inserted immediately
after this column.}

\item{data}{A data frame, containing the new column data, whose length (i.e. number of
columns) must equal that of \code{cols}.}
}
\value{
A \code{patch_insert} object.
}
\description{
S3 class \code{patch_insert} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by inserting one or more new columns.
}
\examples{
colnames(mtcars)
p <- patch_insert("gear", mtcars[2:4])
p <- patch_insert(0L, mtcars[2:4])

# The following are equivalent:
colnames(apply_patch(p, mtcars))
colnames(p(mtcars))

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_insert(22L, mtcars[2:4])
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
