% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_params}
\alias{patch_params}
\title{Get the parameters of a patch object.}
\usage{
patch_params(obj)
}
\arguments{
\item{obj}{A \code{patch} object.}
}
\description{
Returns the \code{patch} parameters object.
}
