% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_perm.R
\name{patch_perm}
\alias{patch_perm}
\title{\code{patch_perm} S3 class constructor}
\usage{
patch_perm(perm)
}
\arguments{
\item{perm}{A vector of integers, being a permutation of 1..n, where n is the
number of columns in the dataset}
}
\value{
A \code{patch_perm} object. When applied to a data set with columns
    a_1, a_2, ..., an, the result is a dataset with columns a_perm[1],
    a_perm[2], ...
}
\description{
S3 class \code{patch_perm} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by an arbitrary permutation of the column indices.
}
\examples{
colnames(mtcars)
p <- patch_perm(c(2L, 3L, 1L))

apply_patch(p, mtcars)

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_cycle(c(1L, 3L, 22L))
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
