% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_rescale.R
\name{patch_rescale}
\alias{patch_rescale}
\title{\code{patch_rescale} S3 class constructor}
\usage{
patch_rescale(cols, shift = 0, scale_factor = 1)
}
\arguments{
\item{cols}{A vector of column identifiers.}

\item{shift}{A numeric scalar. Defaults to zero.}

\item{scale_factor}{A numeric scalar. Defaults to one.}
}
\value{
A \code{patch_rescale} object.
}
\description{
S3 class \code{patch_rescale} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by applying an affine rescaling to numerical data in one or more columns.
}
\examples{
head(mtcars)
p <- patch_rescale(c(1L, 3L), shift = 10, scale_factor = 2)
p

# The following are equivalent:
head(apply_patch(mtcars, p))
head(p(mtcars))

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_rescale(c(1L, 22L), shift = 10, scale_factor = 2)
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
