% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_shift.R
\name{patch_shift}
\alias{patch_shift}
\title{\code{patch_shift} S3 class constructor.}
\usage{
patch_shift(cols, shift)
}
\arguments{
\item{cols}{A vector of column identifiers.}

\item{shift}{A numeric scalar.}
}
\value{
A \code{patch_shift} object.
}
\description{
S3 class \code{patch_shift} which extends the \code{patch} and
\code{function} classes to represent a transformation of a tabular dataset
by shifting (adding a constant to) numerical data in one or more columns.
}
\examples{
head(mtcars)
p <- patch_shift(c(1L, 3L), shift = 2)

# The following are equivalent:
head(apply_patch(mtcars, p))
head(p(mtcars))

# Attempting to apply a patch to an incompatible data frame throws an error.
\dontrun{
p <- patch_shift(c(1L, 22L), shift = 2)
p(mtcars)
}
}
\seealso{
\code{\link{is_valid_columns}} \code{\link{is_compatible_columns}}
}
