% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_type}
\alias{patch_type}
\title{Get the type of a patch object.}
\usage{
patch_type(obj)
}
\arguments{
\item{obj}{A \code{patch} object.}
}
\description{
Returns the \code{patch} type as a character string.
}
