% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\title{Built-in patch types and predicates}
\description{
A collection of standard \code{patch} types and corresponding predicates
available at the package level (i.e. in the package namespace environment,
accessible via the \code{\link{:::}} operator). Each predicate is a function
taking a \code{params} object and returning \code{TRUE} if the form of that
object matches that required to define a patch of the corresponding type.

For instance, a patch of type \code{delete_patch} represents a transformation
which deletes one or more columns from a data frame. The corresponding
predicate tests whether the \code{params} object consists of a vector of
unique negative integers, returning \code{TRUE} in that case and \code{FALSE}
otherwise.
}
\details{
The \code{datadiff:::type_predicates} list provides a mapping between
built-in patch type names and corresponding predicates. This includes the
following patch types:
\itemize{
 \item TYPE: \code{delete_patch}

 ACTION: delete one or more columns from a data frame.

 PREDICATE: \code{params} must be a non-empty vector of unique negative
 integers.

 \item TYPE: \code{permute_patch}

 ACTION: permute the columns of a data frame.

 PREDICATE: \code{params} must be a vector of unique positive integers of
 length two or more.

 \item TYPE: \code{encode_patch}

 ACTION: re-encode categorical data in specified columns.

 PREDICATE: \code{params} must be a list with two elements whose names match
 the character strings contained in \code{datadiff:::COLUMNS} and
 \code{datadiff:::ENCODING}.
 \itemize{
  \item The element named \code{datadiff:::COLUMNS} must be a non-empty
  vector of unique positive integers.
  \item The element named \code{datadiff:::ENCODING} must be a non-empty
  factor containing unique elements with unique names.
  }

 \item TYPE: \code{scale_patch}

 ACTION: rescale numerical data in specified columns.

 PREDICATE: \code{params} must be a list with two elements whose names match
 the character strings contained in \code{datadiff:::COLUMNS} and
 \code{datadiff:::SCALE_FACTOR}.
 \itemize{
  \item The element named \code{datadiff:::COLUMNS} must be a non-empty
  vector of unique positive integers.
  \item The element named \code{datadiff:::SCALE_FACTOR} must be a numerical
  scalar value.
  }
}
}
\seealso{
\code{\link{patch}}
}
