% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_experiment.R
\name{print.synthetic_experiment}
\alias{print.synthetic_experiment}
\title{Print a synthetic experiment object}
\usage{
\method{print}{synthetic_experiment}(x, ...)
}
\arguments{
\item{x}{A \code{synthetic_experiment} object}

\item{...}{Any additional arguments are ignored.}
}
\description{
\code{print} prints its argument and returns it invisibly (via
\code{invisible(x)}).
}
\examples{
config <- configure_synthetic_experiment("mtcars", data_reader = get)
config
# The printed functions names may be called on the config object:
config$get_corruption()

}
\keyword{internal}
