% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{print_patch_params}
\alias{print_patch_params}
\title{Print the parameters associated with a patch object}
\usage{
print_patch_params(patch, df = NULL, indent = "  ", digits = 3,
  n_int = 20, ...)
}
\arguments{
\item{patch}{An elementary \code{patch} object.}

\item{df}{(Optional) A data frame compatible with the given \code{patch}.}

\item{indent}{A character string specifying the intra-patch indentation.}

\item{digits}{The number of decimal places to be used when printing
parameters of type \code{double}. Defaults to 3.}

\item{n_int}{The maximum length at which integer vectors are explicitly printed. Defaults
to 20.}

\item{...}{Additional arguments passed to the \code{param_string} function.}
}
\value{
A character string describing the parameters associated with the
given \code{patch}. If \code{patch} is a composition, the return value is a
vector with one element for each elementary patch obtained by calling
\code{\link{decompose_patch}} on the \code{patch}.
}
\description{
Print the parameters associated with a patch object
}
