% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{render_diff}
\alias{render_diff}
\title{Render a diff in a flexdashboard}
\usage{
render_diff(diff)
}
\arguments{
\item{diff}{Data frame as returned by \link{compare}.}
}
\description{
Render a diff in a flexdashboard
}
