% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.patch.R
\name{return_value.delete.patch}
\alias{return_value.delete.patch}
\title{Generate the patch return value.}
\usage{
\method{return_value}{delete.patch}(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Any additional arguments are ignored.}
}
\description{
Implements the generic function \code{return_value} for a \code{patch} of type \code{delete.patch}. Returns the data frame resulting from the application of the \code{patch} to the given data frame.

In the case of a \emph{delete.patch} this is the given data frame minus those columns whose indices are specified in the patch \code{params} object.
}
\examples{
colnames(mtcars)

p <- patch(-2L)
colnames(return_value(p, mtcars))

}
