% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_patch.R
\name{return_value.permute_patch}
\alias{return_value.permute_patch}
\title{Generate the patch return value.}
\usage{
\method{return_value}{permute_patch}(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Any additional arguments are ignored.}
}
\value{
A patched data frame.
}
\description{
Implements the generic function \code{return_value} for a \code{patch} of
type \code{permute_patch}. Returns the data frame resulting from the
application of the \code{patch} to the given data frame.

In the case of a \emph{permute_patch} this is the given data frame with its
columns permuted according to the vector of indices specified in the patch
\code{params} object, which is interpreted as a \link{cycle}.
}
\examples{
colnames(mtcars)

p <- patch(c(1L, 3L, 4L))
colnames(return_value(p, mtcars))

}
\seealso{
\code{\link{cycle}}
}
