% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{return_value}
\alias{return_value}
\title{Generate the patch return value.}
\usage{
return_value(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Additional arguments.}
}
\description{
Generate the patch return value.
}
