% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_patch.R
\name{return_value.scale_patch}
\alias{return_value.scale_patch}
\title{Generate the patch return value.}
\usage{
\method{return_value}{scale_patch}(obj, df, ...)
}
\arguments{
\item{obj}{A \code{patch} object.}

\item{df}{A data frame.}

\item{...}{Any additional arguments are ignored.}
}
\value{
A patched data frame.
}
\description{
Implements the generic function \code{return_value} for a \code{patch} of
type \code{scale_patch}. Returns the data frame resulting from the
application of the \code{patch} to the given data frame.

In the case of a \emph{scale_patch} this is the given data frame after a
rescaling of the numerical data contained in those columns specified in the
\code{params}, by the specified scale factor.
}
\examples{
purrr::map_dbl(ToothGrowth[c(1, 3)], sum)
p <- patch(scale_patch_params(c(1L, 3L), 2))
purrr::map_dbl(p(ToothGrowth)[c(1, 3)], sum)

}
\seealso{
\code{\link{type.convert}}
}
