% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cols.R
\name{sample_cols}
\alias{sample_cols}
\title{Choose column indices at random}
\usage{
sample_cols(df, condition = is.numeric, exclude_cols = integer(0),
  size = 1)
}
\arguments{
\item{df}{A data frame.}

\item{condition}{A predicate to determine valid candidate columns. An error is thrown if this
predicate returns \code{FALSE} when called on every column in the given data
frame.}

\item{exclude_cols}{An integer vector of column indices to be excluded (i.e. not chosen). An error
is thrown if all of the candidate columns, as identified by the
\code{condition} predicate, are excluded by this argument.}

\item{size}{The number of samples (i.e. column indices) to pick.}
}
\value{
An integer vector of column indices
}
\description{
Choose column indices at random
}
\examples{
# Choose any numeric column:
sample_cols(mtcars, condition = is.numeric)
# Choose a column whose values are binary:
sample_cols(mtcars, condition = function(x) { all(x \%in\% c(0, 1)) })

}
