% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_break.R
\name{sample_patch_break}
\alias{sample_patch_break}
\title{Randomly sample from a distribution of break patches}
\usage{
sample_patch_break(df, rdist = stats::rnorm, colname = NULL,
  condition = function(x) {     TRUE }, exclude_cols = integer(0), seed,
  ...)
}
\arguments{
\item{df}{A data frame}

\item{rdist}{A function for randomly generating the data in the inserted column. Must
contain an argument \code{n} for the number of samples. Defaults to the
standard Normal distribution function \code{rnorm}.}

\item{colname}{The broken column name. If \code{NULL} (the default) the corresponding column
name in the data frame \code{df} is maintained.}

\item{condition}{A predicate which may be used to restrict the break column to those
containing a particular type of data. Defaults to the function which always
returns \code{TRUE} (i.e. no restriction).}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}

\item{...}{Additional arguments passed to the \code{rdist} function.}
}
\description{
Randomly sample from a distribution of break patches
}
\examples{
head(mtcars)
p <- sample_patch_break(mtcars, mean = 10, sd = 4)
p
head(p(mtcars))

# Draw the column data from a given distribution and specify a column name:
p <- sample_patch_break(mtcars, rdist = rexp, rate = 2, colname = "BROKEN")
p
head(p(mtcars))

# Draw the column data from a discrete distribution:
rdist <- function(n, ...) { sample(letters[1:3], size = n, ...) }
p <- sample_patch_break(mtcars, rdist = rdist, replace = TRUE)
p
head(p(mtcars))

}
