% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_delete.R
\name{sample_patch_delete}
\alias{sample_patch_delete}
\title{Pick a delete patch at random}
\usage{
sample_patch_delete(df, exclude_cols = integer(0), seed)
}
\arguments{
\item{df}{A data frame.}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}
}
\description{
Pick a delete patch at random
}
\examples{
sample_patch_delete(mtcars)

}
