% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_identity.R
\name{sample_patch_identity}
\alias{sample_patch_identity}
\title{'Sample' an identity patch}
\usage{
sample_patch_identity(...)
}
\arguments{
\item{...}{Arguments are ignored.}
}
\description{
A trivial helper function (since the identity patch is unique). It is
identical to the \code{patch_identity} constructor except it accepts (and
ignores) any arguments.
}
