% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_insert.R
\name{sample_patch_insert}
\alias{sample_patch_insert}
\title{Randomly sample from a distribution of insert patches}
\usage{
sample_patch_insert(df, rdist = stats::rnorm, colname = "INSERTED",
  exclude_cols = integer(0), seed, ...)
}
\arguments{
\item{df}{A data frame}

\item{rdist}{A function for randomly generating the data in the inserted column. Must
contain an argument \code{n} for the number of samples. Defaults to the
standard Normal distribution function \code{rnorm}.}

\item{colname}{The inserted column name. Defaults to "INSERTED". If a column with that name
already exists, the column will be named "INSERTEDx" where \code{x} is an
integer which increments each time a column with the duplicated name is
inserted.}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the inserted column in the returned patch.}

\item{seed}{A random seed.}

\item{...}{Additional arguments passed to the \code{rdist} function.}
}
\description{
Randomly sample from a distribution of insert patches
}
\examples{
p <- sample_patch_insert(mtcars, mean = 10, sd = 4)
head(p(mtcars))

# Draw the column data from a given distribution:
p <- sample_patch_insert(mtcars, rdist = rexp, rate = 2)
head(p(mtcars))

# Draw the column data from a discrete distribution:
rdist <- function(n, ...) { sample(letters[1:3], size = n, ...) }
p <- sample_patch_insert(mtcars, rdist = rdist, replace = TRUE)
head(p(mtcars))

}
