% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_permute.R
\name{sample_patch_permute}
\alias{sample_patch_permute}
\title{Randomly sample from a distribution of permute patches}
\usage{
sample_patch_permute(df, n = NULL, is_fixed = rep(FALSE, ncol(df)),
  exclude_identity = TRUE, seed, ...)
}
\arguments{
\item{df}{A data frame}

\item{n}{(Optional) An integer specifying the maximum number of columns to be moved
by sample permutations. Defaults to \code{NULL}, which means that all columns
will potentially be moved.}

\item{is_fixed}{A logical vector of length equal to the number of columns in \code{df}. By
default, all elements are \code{FALSE} indicating that any column may
potentially be moved by sample permutations. The \code{is_fixed} argument
is ignored if \code{n} is not \code{NULL}.}

\item{exclude_identity}{A logical flag specifying whether to exclude the identity permutation from
the set of possible sample permutations. Defaults to \code{TRUE}.}

\item{seed}{A random seed.}

\item{...}{Any additional arguments are ignored.}
}
\description{
Randomly sample from a distribution of permute patches
}
\examples{
sample_patch_permute(mtcars)

# Pick a transposition:
sample_patch_permute(mtcars, n = 2L)

# Fix particular columns:
sample_patch_permute(mtcars, is_fixed = c(rep(TRUE, 6), rep(FALSE, 5)))

}
