% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_recode.R
\name{sample_patch_recode}
\alias{sample_patch_recode}
\title{Randomly sample from a distribution of recode patches}
\usage{
sample_patch_recode(df, codes = NULL, condition = function(x) {    
  !(is.double(x) || length(unique(x)) > 2^5) },
  exclude_cols = integer(0), seed, ...)
}
\arguments{
\item{df}{A data frame}

\item{codes}{A vector of codes. If \code{NULL} (the default), \code{codes} is a minimal
vector of integers coerced to character type.}

\item{condition}{A predicate which may be used to restrict the recode column to those
containing a particular type of data. Defaults to the negation of
(\code{is.double} or contains more than 2^5 unique values).}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}

\item{...}{Additional arguments passed to the \code{sample} function.}
}
\value{
A recode patch with randomly sampled parameters.
}
\description{
Randomly sample from a distribution of recode patches
}
\examples{
condition <- function(x) { all(as.integer(x) == x) }
p <- sample_patch_recode(mtcars, codes = letters[1:26], condition = condition)
head(p(mtcars))

condition <- function(x) { length(unique(x)) == 2 }
p <- sample_patch_recode(mtcars, codes = c(TRUE, FALSE), condition = condition)
head(p(mtcars))

}
