% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_rescale.R
\name{sample_patch_rescale}
\alias{sample_patch_rescale}
\title{Randomly sample from a distribution of rescale patches}
\usage{
sample_patch_rescale(df, rdist_shift = purrr::partial(stats::rnorm, mean =
  mean_shift), mean_shift = 0, relative_shift = NULL,
  rdist_scale = purrr::partial(stats::rnorm, mean = mean_scale),
  mean_scale = 1, exclude_cols = integer(0), seed)
}
\arguments{
\item{df}{A data frame}

\item{rdist_shift}{A function for randomly generating the shift. Must contain an argument
\code{n} for the number of samples. Defaults to the Normal distribution
function \code{rnorm} with zero mean and standard deviation.}

\item{mean_shift}{(Optional) A number specifying the \code{mean} parameter in the Normal
distribution function which is the default value for \code{rdist_shift}. Note
that this argument is ignored unless \code{rdist_shift} takes its default
value. Defaults to 0.}

\item{relative_shift}{(Optional) A number specifying a relative shift size. If non-\code{NULL}, the
\code{rdist} and \code{mean_shift} arguments are ignored and the shift is
sampled from the Normal distribution with mean (and standard deviation,
resp.) given by \code{relative_shift} multiplied by the median
(resp. standard deviation) of the values in the selected column
of \code{df} (ignoring \code{NA}s). The \code{relative_shift} argument defaults
to \code{NULL} to produce a shift which is absolute, not relative.}

\item{rdist_scale}{A function for randomly generating the scale parameter. Must contain an
argument \code{n} for the number of samples. Defaults to the standard Normal
distribution function \code{rnorm}.}

\item{mean_scale}{(Optional) A number specifying the \code{mean} parameter in the Normal
distribution function which is the default value for \code{rdist_scale}. Note
that this argument is ignored unless \code{rdist_scale} takes its default
value. Defaults to 1.}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}
}
\value{
A rescale patch with randomly sampled parameters.
}
\description{
Randomly sample from a distribution of rescale patches
}
\examples{

# By default, parameter values are sampled from a standard Normal distribution:
sample_patch_rescale(mtcars)
sample_patch_rescale(mtcars, mean_shift = 10, mean_scale = 2)

}
