% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_scale.R
\name{sample_patch_scale}
\alias{sample_patch_scale}
\title{Randomly sample from a distribution of scale patches}
\usage{
sample_patch_scale(df, rdist = purrr::partial(stats::rnorm, mean = mean),
  mean = 1, exclude_cols = integer(0), seed, ...)
}
\arguments{
\item{df}{A data frame}

\item{rdist}{A function for randomly generating the scale factor. Must
contain an argument \code{n} for the number of samples. Defaults to the Normal
distribution function \code{rnorm} with unit mean and standard deviation.}

\item{mean}{(Optional) A number specifying the \code{mean} parameter in the Normal
distribution function which is the default value for \code{rdist}. Note that
this argument is ignored unless \code{rdist} takes its default value. Defaults
to 1.}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}

\item{...}{Additional arguments passed to the \code{rdist} function.}
}
\value{
A scale patch with randomly sampled parameters.
}
\description{
Randomly sample from a distribution of scale patches
}
\examples{
# By default, the scale parameter is sampled from a standard Normal distribution:
sample_patch_scale(mtcars)
sample_patch_scale(mtcars, mean = 2)

}
