% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_shift.R
\name{sample_patch_shift}
\alias{sample_patch_shift}
\title{Randomly sample from a distribution of shift patches}
\usage{
sample_patch_shift(df, rdist = stats::rnorm, relative_shift = NULL,
  exclude_cols = integer(0), seed, ...)
}
\arguments{
\item{df}{A data frame}

\item{rdist}{A function for randomly generating the shift parameter. Must
contain an argument \code{n} for the number of samples. Defaults to the
standard Normal distribution function \code{rnorm}.}

\item{relative_shift}{(Optional) A number specifying a relative shift size. If non-\code{NULL}, the
\code{rdist} argument is ignored and the shift is sampled from the Normal
distribution with mean (and standard deviation, respectively) given by
\eqn{relative_shift} multiplied by the median (respectively standard deviation)
of the values in the selected column of \code{df} (ignoring NAs). The
\code{relative_shift} argument defaults to \code{NULL} to produce a shift
which is absolute, not relative, and is sampled using the \code{rdist}
function.}

\item{exclude_cols}{An integer vector of column indices to be excluded from the set of possible
target columns for the returned patch.}

\item{seed}{A random seed.}

\item{...}{Additional arguments passed to the \code{rdist} function.}
}
\value{
A shift patch with randomly sampled parameters.
}
\description{
Randomly sample from a distribution of shift patches
}
\examples{
# By default, the shift parameter is sampled from a standard Normal distribution:
sample_patch_shift(mtcars)
sample_patch_shift(mtcars, mean = 2)

}
