% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.patch.R
\name{scale_patch_params}
\alias{scale_patch_params}
\title{Construct parameters for the \code{scale.patch} type}
\usage{
scale_patch_params(column_indices = integer(), scale_factor = numeric())
}
\arguments{
\item{column_indices}{An integer vector of column indices}

\item{scale_factor}{A numerical scalar specifying a scale factor.}
}
\description{
Converts a vector of integer column indices and a numerical scale factor
into a \code{params} object suitable for constructing a patch of type
\code{scale.patch}.
}
\seealso{
\code{\link{patch}}
}
