% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_pairwise_assignment.R
\name{solve_pairwise_assignment}
\alias{solve_pairwise_assignment}
\title{Solve the linear sum assignment problem}
\usage{
solve_pairwise_assignment(m, maximum = FALSE)
}
\arguments{
\item{m}{A matrix with nonnegative entries.}

\item{maximum}{A logical flag indicating whether to minimize of maximize the sum of assigned
costs.}
}
\value{
An integer vector.
}
\description{
Helper function for \code{\link{ddiff}}. Solves the problem of optimally
assigning column... TODO.
}
