% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split a data frame into two pieces}
\usage{
split_data(df, split)
}
\arguments{
\item{df}{A data frame.}

\item{split}{A number in the unit interval specifying the splitting ratio.}
}
\value{
A list containing two data frames.
}
\description{
Split a data frame into two pieces
}
\examples{
split_data(head(mtcars), split = 1/2)

}
