% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{tv}
\alias{tv}
\title{Total variation distance for two discrete variables (as factors)}
\usage{
tv(v1, v2)
}
\arguments{
\item{v1, f2}{A pair of factors. Both must have at least one non-missing value.}
}
\value{
A number between 0 and 1 inclusive.
}
\description{
Compute the total variation distance for samples from two discrete
distributions, coded as unordered factors. The result is a number in the
(closed) unit interval.
}
