% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{apply_patch}
\alias{apply_patch}
\title{Apply a patch to a data frame.}
\usage{
apply_patch(df, patch, ...)
}
\arguments{
\item{df}{A data frame.}

\item{patch}{A patch object.}

\item{...}{Additional arguments passed to the patch function.}
}
\value{
A transformed data frame.
}
\description{
Transform one data frame into another by applying a patch.
}
