% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnwise_candidates.R
\name{columnwise_candidates}
\alias{columnwise_candidates}
\title{Generate candidate patches by columnwise comparison}
\usage{
columnwise_candidates(df1, df2, mismatch = purrr::partial(diffness, scale =
  FALSE), patch_generators = list(gen_patch_transform),
  patch_penalties = 0.2, break_penalty = 0.95,
  scale_penalty = scale_penalty, verbose = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{mismatch}{Mismatch method. The default is (unscaled) \code{\link{diffness}}.}

\item{patch_generators}{A list of patch generator functions.}

\item{patch_penalties}{A vector of patch penalties corresponding to the \code{patch_generators}
list. The lengths of these two parameters must be equal.}

\item{verbose}{A logical flag.}
}
\value{
A nested list of patch objects with two numeric attributes:
\code{mismatch} and \code{penalty}.
}
\description{
TODO.
}
