% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnwise_candidates.R
\name{columnwise_candidates}
\alias{columnwise_candidates}
\title{Generate candidate patches by columnwise comparison}
\usage{
columnwise_candidates(df1, df2, mismatch, patch_generators, patch_penalties,
  break_penalty, penalty_scaling, mismatch_attr = "mismatch",
  penalty_attr = "penalty", verbose = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{mismatch}{Mismatch method. The default is \code{\link{diffness}}.}

\item{patch_generators}{A list of patch generator functions from which candidate patches will be
generated for every pair of columns (one each from \code{df1} & \code{df2}).}

\item{patch_penalties}{A numeric vector of patch penalties corresponding to the \code{patch_generators}
list. The lengths of these two arguments must be equal.}

\item{break_penalty}{The penalty associated with a break patch. No scaling is applied to the
\code{break_penalty}.}

\item{penalty_scaling}{A function to be used to scale the penalty associated with each patch.
Defaults to \code{\link{ks_scaling}}.}

\item{mismatch_attr}{The name of the attribute in the return value containing the calculated mismatch associated with
each candidate patch. Defaults to "mismatch".}

\item{penalty_attr}{The name of the attributein the return value containing the calculated penalty associated with
each candidate patch. Defaults to "penalty".}

\item{verbose}{A logical flag.}
}
\value{
A nested list of patch objects, each with two numeric attributes:
one for the residual columnwise mismatch and one for the patch penalty.
}
\description{
Helper function for \code{\link{ddiff}}. Generates a candidate patch for each
pair of columns (one each from \code{df1} & \code{df2}) and returns them in a
nested list.

Possible candidates are:
\itemize{
  \item the identity patch
  \item a patch generated from one of the given \code{patch_generators}
  \item a break patch.
}
The "total cost" of each of these possible candidates is computed as the
sum of the residual (columwise) mismatch (after application of the patch)
plus the (scaled) penalty associated with the patch, except for the case of
the break patch for which the mismatch is always taken to be zero and the
penalty is not scaled.

The actual candidate is then the one with the minimum total cost.
}
\details{
Each element in the \code{patch_generators} list must be a patch generator
function. That is, a function, with arguments \code{df1}, \code{df2},
\code{mismatch}, \code{col1} and \code{col2}, which returns a patch such that
the mismatch between column \code{col2} in \code{df2} and the patched column
\code{col1} in \code{df1} is minimised to the extent possible using the type
of patch in question. If the attempt at patch generation fails, the generator
function must return \code{NULL}. An example of such a patch generator is
\code{\link{gen_patch_transform}}.
}
