% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_data}
\alias{compare_data}
\title{Compare two data frames}
\usage{
compare_data(
  x,
  y,
  context_rows = c(3L, 3L),
  context_cols = everything(),
  max_differences = Inf,
  tolerance = .Machine$double.eps^0.5
)
}
\arguments{
\item{x, y}{Data frames to compare.}

\item{context_rows}{Integer vector of length two indicating the number of context
rows to include before and after a difference row.}

\item{context_cols}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns to include as context.}

\item{max_differences}{Maximum number of differences to return.}

\item{tolerance}{Numeric tolerance for comparing numeric values.}
}
\value{
A data frame containing differences between \code{x} and \code{y} with the
following columns:
\itemize{
\item \code{.row} - The row number from the original data frames
\item \code{.join_type} - Whether the row is in \code{"x"}, \code{"y"}, or \code{"both"}
\item \code{.diff_type} - Whether the row is a \code{"diff"} or \code{"context"} row
\item \code{.source} - For diff rows, whether this is the \code{"x"} or \code{"y"} version;
\code{NA} for context rows
}
Plus the original data columns (context columns and columns with differences).
}
\description{
Compare two data frames
}
