% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_diff}
\alias{compare_diff}
\title{Diff data frames that have been compare joined}
\usage{
compare_diff(data, context = c(3L, 3L), max_differences = Inf)
}
\arguments{
\item{data}{Data frame as returned by \link{compare_join}.}

\item{context}{Integer vector of length two indicating the number of context
row to include before and after a difference row.}

\item{max_differences}{Maximum number of differences to return.}
}
\description{
Diff data frames that have been compare joined
}
