% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff.R
\name{ddiff}
\alias{ddiff}
\title{Diff two data frames}
\usage{
ddiff(df1, df2, mismatch = purrr::partial(diffness, scale = FALSE),
  patch_generators = list(gen_patch_transform), patch_penalties = 0.6,
  break_penalty = 0.99, perm_penalty = 0.4, penalty_scaling = ks_scaling,
  as.list = FALSE, verbose = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{as.list}{A logical flag. If \code{TRUE} the return value is a list of patches.
Otherwise a composite patch object is returned.}

\item{verbose}{A logical flag.}
}
\value{
A list of patch objects, or their composition if \code{composed} is
\code{TRUE}.
}
\description{
TODO.
}
