% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff.R
\name{ddiff}
\alias{ddiff}
\title{Diff two data frames}
\usage{
ddiff(df1, df2, cost_permute, cost_transform, cost_break = 0.99,
  as.list = FALSE)
}
\arguments{
\item{df1, df2}{A pair of data frames.}

\item{cost_permute}{A number in the unit interval representing the cost of a permute patch
operation.}

\item{cost_transform}{A number in the unit interval representing the cost of a transformation patch
(i.e. affine for continuous data, recode for categorical) operation.}

\item{cost_break}{A number in the unit interval representing the cost of a break patch
operation.}

\item{as.list}{A logical flag. If \code{TRUE} the return value is a list of patches.
Otherwise a composite patch object is returned.}
}
\value{
A list of patch objects, or their composition if \code{composed} is
\code{TRUE}.
}
\description{
TODO.
}
