% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{decompose_patch}
\alias{decompose_patch}
\title{Decompose a composition of patches}
\usage{
decompose_patch(patch)
}
\arguments{
\item{patch}{A composition of \code{patch} objects.}
}
\value{
A list of elementary \code{patch} objects.
}
\description{
Returns the elementary \code{patch} objects in a list in the order in which
they are applied in the composition.
}
