% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{decompose_patch}
\alias{decompose_patch}
\title{Decompose a composition of patches}
\usage{
decompose_patch(patch)
}
\arguments{
\item{patch}{A composition of \code{patch} objects.}
}
\value{
A list of elementary \code{patch} objects.
}
\description{
Patches may be composed using the \code{\link{compose}} function and the
result is a composite patch object. Given such an object, this function
returns its constituent elementary \code{patch} objects in a list in the
order in which they are applied in the composition. Compositions of
compositions are handled via tree recursion.
}
\examples{
patch <- compose_patch(patch_permute(8:1), patch_shift(2L, shift = 10))
decompose_patch(patch)

}
\seealso{
\code{\link{compose_patch}}
}
