% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdata.R
\name{diffdata}
\alias{diffdata}
\title{Diff Data Frames}
\usage{
diffdata(x, y, max_differences = 10, context = c(3L, 3L))
}
\arguments{
\item{x, y}{Data frames to diff.}

\item{max_differences}{Maximum differences to detect.}

\item{context}{Number of observations before and after difference rows to
contextualize differences.}
}
\description{
Diff Data Frames
}
