% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.double}
\alias{diffness.double}
\title{Compute the mismatch between two continuous numeric vectors}
\usage{
\method{diffness}{double}(x, y, diff = ks, ...)
}
\arguments{
\item{x, y}{A pair of vectors of type \code{double}.}

\item{diff}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for continuous
numeric data.}

\item{...}{Additional arguments are ignored.}
}
\description{
Vectors of type \code{double} are treated as continuous numerical data.
}
\details{
The mismatch between vectors of different types is always 1.
}
