% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness.integer}
\alias{diffness.integer}
\title{Compute the mismatch between two integer vectors}
\usage{
\method{diffness}{integer}(x, y, scale = TRUE, diff = ks, ...)
}
\arguments{
\item{x, y}{A pair of vectors of type \code{integer}.}

\item{scale}{A logical flag. If \code{TRUE} (the default) the result is scaled by the
length of the data.}

\item{diff}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov) for integer
data.}

\item{...}{Additional arguments are ignored.}
}
\description{
By default, integer vectors are treated as ordered categorical data. To
ignore ordering, set the \code{diff} argument to \code{tv} (total variation
distance).
}
\details{
The mismatch between an integer vector and a non-integer numeric vector is
always +Inf.
}
