% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffness.R
\name{diffness}
\alias{diffness}
\title{Compute the mismatch between two datasets}
\usage{
diffness(x, y, ...)
}
\arguments{
\item{x, y}{A pair of datasets}

\item{...}{Additional arguments passed to methods.}
}
\value{
A non-negative number quantifying the mismatch between the two
datasets.
}
\description{
\code{diffness} is a generic function: the type of the first argument
determines which method is called.
}
\examples{
diffness(purrr::rbernoulli(10), purrr::rbernoulli(10))
diffness(1:10, 10:1)
diffness(1:10, 20:1)
diffness(stats::rnorm(10), stats::rnorm(10, sd = 2))
diffness(letters[1:10], letters[20:1])
diffness(mtcars, mtcars[1:10, ])
}
