% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_affine.R
\name{gen_patch_affine}
\alias{gen_patch_affine}
\title{Generate an affine patch}
\usage{
gen_patch_affine(df, cols, y, robust = TRUE, ...)
}
\arguments{
\item{df}{A data frame. The column specified in the \code{cols} argument must contain a
vector of type \code{double} with at least two non-missing values.}

\item{cols}{A column identifier (integer or string column name) with length 1.}

\item{y}{A 'target' vector of type \code{double} with least two non-missing values.}

\item{robust}{A logical flag. If \code{TRUE} (the default) the median absolute deviation is
used as a robust measure of statistical dispersion. Otherwise the standard
deviation is used.}

\item{...}{Additional arguments passed to the .}
}
\value{
A composite \code{patch} object.
}
\description{
Generates a composite 'affine' \code{patch} object, formed as the composition
of a \code{patch_shift} and a \code{patch_scale}, whose 'shift' and 'scale
factor' parameters hav been selected with the aim of minimising the mismatch
between the vector \code{y} and the patch applied to the given data frame.
}
\seealso{
\code{\link{gen_patch_shift}} \code{\link{gen_patch_scale}}
}
