% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_scale.R
\name{gen_patch_scale}
\alias{gen_patch_scale}
\title{Generate a scale patch}
\usage{
gen_patch_scale(df1, col1, df2, col2 = col1, robust = TRUE, ...)
}
\arguments{
\item{df1}{A data frame. The column specified in the \code{col1} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{col1}{A column identifier (integer or string column name) with length 1.}

\item{df2}{A data frame. The column specified in the \code{col2} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{col2}{A column identifier (integer or string column name) with length 1. By default
this takes the value of \code{col1}.}

\item{robust}{A logical flag. If \code{TRUE} (the default) the median absolute deviation is
used as a robust measure of statistical dispersion. Otherwise the standard
deviation is used.}

\item{...}{Additional arguments passed to the function used to calculate the statistical
dispersion (i.e. \code{\link{mad}} if \code{robust} is \code{TRUE}, otherwise
\code{\link{sd}}).}
}
\value{
A \code{patch_scale} object.
}
\description{
Generates a \code{patch_scale} object whose 'scale factor' parameter has been
selected with the aim of minimising the mismatch between the specified
columns after application of the patch to \code{df1}.
}
\seealso{
\code{\link{gen_patch_affine}} \code{\link{mad}}
}
