% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_scale.R
\name{gen_patch_scale}
\alias{gen_patch_scale}
\title{Generate a scale patch}
\usage{
gen_patch_scale(df1, df2, mismatch = ks, col1, col2 = col1, ...)
}
\arguments{
\item{df1}{A data frame. The column specified in the \code{col1} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{df2}{A data frame. The column specified in the \code{col2} argument must contain
a vector of type \code{double} with at least two non-missing values.}

\item{mismatch}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov).}

\item{col1}{A column identifier (integer or string column name) of length 1.}

\item{col2}{A column identifier (integer or string column name) of length 1. By default
this takes the value of \code{col1}.}
}
\value{
A \code{patch_scale} object.
}
\description{
Generates a \code{patch_scale} object whose 'scale factor' parameter has been
selected with the aim of minimising the mismatch between the specified
columns after application of the patch to \code{df1}.
}
\seealso{
\code{\link{gen_patch_affine}} \code{\link{mad}}
}
