% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patch_shift.R
\name{gen_patch_shift}
\alias{gen_patch_shift}
\title{Generate a shift patch}
\usage{
gen_patch_shift(df1, df2, mismatch = ks, col1, col2 = col1)
}
\arguments{
\item{df1}{A data frame. The column specified in the \code{col1} argument must contain
a vector of type \code{double} with at least one non-missing value.}

\item{df2}{A data frame. The column specified in the \code{col2} argument must contain
a vector of type \code{double} with at least one non-missing value.}

\item{mismatch}{Mismatch method. The default is \code{ks} (Kolmogorov-Smirnov).}

\item{col1}{A column identifier (integer or string column name) of length 1.}

\item{col2}{A column identifier (integer or string column name) of length 1. By default
this takes the value of \code{col1}.}
}
\value{
A \code{patch_shift} object.
}
\description{
Generates a \code{patch_shift} object whose \code{shift} parameter has been
selected with the aim of minimising the mismatch between the specified
columns after application of the patch to \code{df1}.
}
\details{
Uses the \code{\link{optim}} function to optimise the \code{shift}
parameter for the given mismatch method. If the numerical optimisation fails
to converge an error is thrown.
}
\examples{
gen_patch_shift(mtcars, mtcars, col1 = "wt", col2 = "qsec")

}
\seealso{
\code{\link{gen_patch_affine}}
}
