% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{get_patch_params}
\alias{get_patch_params}
\title{Get the parameters associated with a patch object.}
\usage{
get_patch_params(patch)
}
\arguments{
\item{patch}{A \code{patch} object.}
}
\value{
An object containing the parameters associated with the given
\code{patch}
}
\description{
Returns the \code{patch} parameters object.
}
