% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heartdisease}
\alias{heartdisease}
\title{Heart Disease data set}
\format{A data frame with 303 rows and 14 variables.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
\usage{
heartdisease
}
\description{
Heart Disease data set from the UCI Machine Learning Repository.
}
\keyword{datasets}
