% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is_patch}
\alias{is_patch}
\title{Test for a patch object}
\usage{
is_patch(obj, allow_composed = TRUE)
}
\arguments{
\item{obj}{Any object.}

\item{allow_composed}{A logicial flag. If \code{TRUE} (the default) then compositions of patches
(constructed using either the \code{compose_patch} function or the
\code{compose} function from the \code{purrr} package) are considered to be
patches. Otherwise, only elementary patch objects are considered to be
patches.}
}
\description{
Returns \code{TRUE} if the \code{obj} argument is a \code{patch} object.
}
\examples{
is_patch(patch_rescale(1L, shift = 0, scale_factor = 2))
is_patch(patch_identity())
is_patch(patch_identity)
is_patch(mtcars)
is_patch(compose_patch(patch_permute(10:1), patch_delete(1L)))
is_patch(compose_patch(patch_permute(10:1), patch_delete(1L)),
         allow_composed = FALSE)

}
