% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is_patch}
\alias{is_patch}
\title{Test for a patch object}
\usage{
is_patch(obj)
}
\arguments{
\item{obj}{Any object.}
}
\description{
Returns \code{TRUE} if the \code{obj} argument is a \code{patch} object.
}
