% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_scaling.R
\name{ks_scaling}
\alias{ks_scaling}
\title{Scale a penalty for the two-sample Kolmogorov-Smirnov test}
\usage{
ks_scaling(penalty, nx, ny)
}
\arguments{
\item{penalty}{A numeric penalty}

\item{nx, ny}{The lengths of the vectors}
}
\value{
A (numeric) scaled penalty.
}
\description{
A penalty is a number associated with a \code{patch} which represents the
cost applying the patch. Only patches which reduce the mismatch by an amount
in excess of their associated cost are deemed worthwhile. However, the
penalty is a fixed number (in the unit interval) whereas the mismatch
typically scales with the size of the dataset. Hence, for comparability, the
penalty must also be scaled.

This function scales the penalty in proportion to the inverse square root of
the sample size, so that small mismatches count more the more data
points there are.
Specifically, the scaling is that appropriate for the
\href{https://en.wikipedia.org/wiki/Kolmogorov-Smirnov_test}{two-sample
Kolmogorov-Smirnov test} where the confidence threshold \eqn{alpha}
corresponds to one minus the penalty.

If the \code{penalty} is zero, the return value is also zero.
}
\examples{
ks_scaling(0.5, nx = 100, ny = 100)

}
\seealso{
\code{\link{linear_scaling}}
}
