% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_identity.R
\name{patch_identity}
\alias{patch_identity}
\title{\code{patch_identity} S3 class constructor}
\usage{
patch_identity()
}
\value{
A \code{patch_identity} object.
}
\description{
S3 class \code{patch_identity} which extends the \code{patch} and
\code{function} classes to represent the identity transformation of a
tabular dataset.
}
\examples{
p <- patch_identity()

}
