% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_type}
\alias{patch_type}
\title{Get the patch type.}
\usage{
patch_type(patch, short = TRUE)
}
\arguments{
\item{patch}{A patch object.}

\item{short}{A logical flag. If \code{TRUE} (the default) the prefix "patch_" is omitted
from the return value.}
}
\value{
A character string, or a vector if \code{patch} is a composition of
patches.
}
\description{
Transform one data frame into another by applying a patch.
}
