% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_type}
\alias{patch_type}
\title{Get the patch type.}
\usage{
patch_type(patch, short = TRUE, unique = FALSE)
}
\arguments{
\item{patch}{A patch object.}

\item{short}{A logical flag. If \code{TRUE} (the default) the prefix "patch_" is omitted
from the return value.}

\item{unique}{A logical flag. If \code{TRUE} any duplicates in the return value will be
removed. This may happen only if \code{patch} is a composition. Defaults to
\code{FALSE}.}
}
\value{
A character string, or a vector if \code{patch} is a composition of
patches.
}
\description{
Get the patch type.
}
\examples{
patch <- compose_patch(patch_shift(6L, shift = 4),
                       patch_permute(8:1),
                       patch_shift(2L, shift = 10))
patch_type(patch)
patch_type(patch, unique = TRUE)
}
