% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{print.patch}
\alias{print.patch}
\title{Print a patch object.}
\usage{
\method{print}{patch}(x, ...)
}
\arguments{
\item{x}{A patch object}

\item{...}{Any additional arguments are ignored.}
}
\description{
\code{print} prints its argument and returns it invisibly (via
\code{invisible(x)}).
}
\details{
To print a composition of patches constructed using the \code{\link{compose}}
use \code{\link{decompose_patch}}.
}
\seealso{
\code{\link{decompose_patch}}
}
