% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{print_patch_params}
\alias{print_patch_params}
\title{Print the parameters associated with a patch object.}
\usage{
print_patch_params(patch, digits = 3)
}
\arguments{
\item{patch}{A \code{patch} object.}

\item{digits}{Integer indicating the number of decimal places to be used when printing
parameters of type \code{double}. Defaults to 3.}
}
\value{
A character string describing the parameters associated with the given
\code{patch}.
}
\description{
Print the parameters associated with a patch object.
}
